% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apixbrl_company.R
\docType{class}
\name{APIXBRLCompany}
\alias{APIXBRLCompany}
\title{APIXBRLCompany}
\format{
An \code{R6Class} generator object
}
\description{
APIXBRLCompany Class

APIXBRLCompany Class
}
\details{
Create a new APIXBRLCompany
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entity_name}}{character \link{optional}}

\item{\code{entity_id}}{integer \link{optional}}

\item{\code{ticker}}{character \link{optional}}

\item{\code{entity_code}}{character \link{optional}}

\item{\code{sic_code}}{integer \link{optional}}

\item{\code{naics}}{integer \link{optional}}

\item{\code{naics2}}{integer \link{optional}}

\item{\code{assets}}{numeric \link{optional}}

\item{\code{revenue}}{numeric \link{optional}}

\item{\code{state}}{character \link{optional}}

\item{\code{country}}{character \link{optional}}

\item{\code{zip}}{character \link{optional}}

\item{\code{EntityIncorporationStateCountryCode}}{character \link{optional}}

\item{\code{EntityTaxIdentificationNumber}}{character \link{optional}}

\item{\code{SecurityExchangeName}}{character \link{optional}}

\item{\code{FYEndDateChange}}{integer \link{optional}}

\item{\code{status}}{character \link{optional}}

\item{\code{Street1}}{character \link{optional}}

\item{\code{Street2}}{character \link{optional}}

\item{\code{most_recent_filing}}{character \link{optional}}

\item{\code{most_recent_full_year_end}}{character \link{optional}}

\item{\code{most_recent_fiscal_year}}{integer \link{optional}}

\item{\code{most_recent_complete_calendar_year}}{integer \link{optional}}

\item{\code{most_recent_complete_fiscal_year}}{integer \link{optional}}

\item{\code{most_recent_filing_calendar_period}}{integer \link{optional}}

\item{\code{most_recent_filing_calendar_year}}{integer \link{optional}}

\item{\code{first_filing}}{character \link{optional}}

\item{\code{cik_code}}{object \link{optional}}

\item{\code{SICCategory}}{character \link{optional}}

\item{\code{naics_code}}{object \link{optional}}

\item{\code{city}}{character \link{optional}}

\item{\code{street_address}}{character \link{optional}}

\item{\code{naics_minor_group_title}}{character \link{optional}}

\item{\code{SICGroupMinorGroupTitle}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-APIXBRLCompany-new}{\code{APIXBRLCompany$new()}}
\item \href{#method-APIXBRLCompany-toJSON}{\code{APIXBRLCompany$toJSON()}}
\item \href{#method-APIXBRLCompany-fromJSON}{\code{APIXBRLCompany$fromJSON()}}
\item \href{#method-APIXBRLCompany-toJSONString}{\code{APIXBRLCompany$toJSONString()}}
\item \href{#method-APIXBRLCompany-fromJSONString}{\code{APIXBRLCompany$fromJSONString()}}
\item \href{#method-APIXBRLCompany-validateJSON}{\code{APIXBRLCompany$validateJSON()}}
\item \href{#method-APIXBRLCompany-toString}{\code{APIXBRLCompany$toString()}}
\item \href{#method-APIXBRLCompany-isValid}{\code{APIXBRLCompany$isValid()}}
\item \href{#method-APIXBRLCompany-getInvalidFields}{\code{APIXBRLCompany$getInvalidFields()}}
\item \href{#method-APIXBRLCompany-print}{\code{APIXBRLCompany$print()}}
\item \href{#method-APIXBRLCompany-clone}{\code{APIXBRLCompany$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIXBRLCompany-new"></a>}}
\if{latex}{\out{\hypertarget{method-APIXBRLCompany-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new APIXBRLCompany class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIXBRLCompany$new(
  entity_name = NULL,
  entity_id = NULL,
  ticker = NULL,
  entity_code = NULL,
  sic_code = NULL,
  naics = NULL,
  naics2 = NULL,
  assets = NULL,
  revenue = NULL,
  state = NULL,
  country = NULL,
  zip = NULL,
  EntityIncorporationStateCountryCode = NULL,
  EntityTaxIdentificationNumber = NULL,
  SecurityExchangeName = NULL,
  FYEndDateChange = NULL,
  status = NULL,
  Street1 = NULL,
  Street2 = NULL,
  most_recent_filing = NULL,
  most_recent_full_year_end = NULL,
  most_recent_fiscal_year = NULL,
  most_recent_complete_calendar_year = NULL,
  most_recent_complete_fiscal_year = NULL,
  most_recent_filing_calendar_period = NULL,
  most_recent_filing_calendar_year = NULL,
  first_filing = NULL,
  cik_code = NULL,
  SICCategory = NULL,
  naics_code = NULL,
  city = NULL,
  street_address = NULL,
  naics_minor_group_title = NULL,
  SICGroupMinorGroupTitle = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity_name}}{entity_name}

\item{\code{entity_id}}{entity_id}

\item{\code{ticker}}{ticker}

\item{\code{entity_code}}{entity_code}

\item{\code{sic_code}}{sic_code}

\item{\code{naics}}{naics}

\item{\code{naics2}}{naics2}

\item{\code{assets}}{assets}

\item{\code{revenue}}{revenue}

\item{\code{state}}{state}

\item{\code{country}}{country}

\item{\code{zip}}{zip}

\item{\code{EntityIncorporationStateCountryCode}}{EntityIncorporationStateCountryCode}

\item{\code{EntityTaxIdentificationNumber}}{EntityTaxIdentificationNumber}

\item{\code{SecurityExchangeName}}{SecurityExchangeName}

\item{\code{FYEndDateChange}}{FYEndDateChange}

\item{\code{status}}{status}

\item{\code{Street1}}{Street1}

\item{\code{Street2}}{Street2}

\item{\code{most_recent_filing}}{most_recent_filing}

\item{\code{most_recent_full_year_end}}{most_recent_full_year_end}

\item{\code{most_recent_fiscal_year}}{most_recent_fiscal_year}

\item{\code{most_recent_complete_calendar_year}}{most_recent_complete_calendar_year}

\item{\code{most_recent_complete_fiscal_year}}{most_recent_complete_fiscal_year}

\item{\code{most_recent_filing_calendar_period}}{most_recent_filing_calendar_period}

\item{\code{most_recent_filing_calendar_year}}{most_recent_filing_calendar_year}

\item{\code{first_filing}}{first_filing}

\item{\code{cik_code}}{cik_code}

\item{\code{SICCategory}}{SICCategory}

\item{\code{naics_code}}{naics_code}

\item{\code{city}}{city}

\item{\code{street_address}}{street_address}

\item{\code{naics_minor_group_title}}{naics_minor_group_title}

\item{\code{SICGroupMinorGroupTitle}}{SICGroupMinorGroupTitle}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIXBRLCompany-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-APIXBRLCompany-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIXBRLCompany$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
APIXBRLCompany in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIXBRLCompany-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-APIXBRLCompany-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of APIXBRLCompany
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIXBRLCompany$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of APIXBRLCompany
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIXBRLCompany-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-APIXBRLCompany-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIXBRLCompany$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
APIXBRLCompany in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIXBRLCompany-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-APIXBRLCompany-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of APIXBRLCompany
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIXBRLCompany$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of APIXBRLCompany
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIXBRLCompany-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-APIXBRLCompany-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to APIXBRLCompany and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIXBRLCompany$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIXBRLCompany-toString"></a>}}
\if{latex}{\out{\hypertarget{method-APIXBRLCompany-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIXBRLCompany$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of APIXBRLCompany
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIXBRLCompany-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-APIXBRLCompany-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIXBRLCompany$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIXBRLCompany-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-APIXBRLCompany-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIXBRLCompany$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIXBRLCompany-print"></a>}}
\if{latex}{\out{\hypertarget{method-APIXBRLCompany-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIXBRLCompany$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIXBRLCompany-clone"></a>}}
\if{latex}{\out{\hypertarget{method-APIXBRLCompany-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIXBRLCompany$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
