% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_metrics_api.R
\docType{class}
\name{AvailableMetricsApi}
\alias{AvailableMetricsApi}
\title{AvailableMetrics operations}
\format{
An \code{R6Class} generator object
}
\description{
AvailableMetricsApi
}
\details{
Calcbench API

No description provided (generated by Openapi Generator https://github.com/calcbenchRtools/calcbenchR-generator)

The version of the OpenAPI document: v1
Generated by: https://calcbenchR-generator.tech
}
\section{Methods}{

\describe{
\strong{ AvailableMetricsGetValues } \emph{  }

\itemize{
\item \emph{ @returnType } \link{MetricGroups} \cr

\item status code : 200 | OK

\item return type : MetricGroups
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  AvailableMetricsGetValues  ####################

library(calcbenchR)

api_instance <- AvailableMetricsApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$AvailableMetricsGetValues(data_file = "result.txt")
result <- api_instance$AvailableMetricsGetValues()
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AvailableMetricsApi-new}{\code{AvailableMetricsApi$new()}}
\item \href{#method-AvailableMetricsApi-AvailableMetricsGetValues}{\code{AvailableMetricsApi$AvailableMetricsGetValues()}}
\item \href{#method-AvailableMetricsApi-AvailableMetricsGetValuesWithHttpInfo}{\code{AvailableMetricsApi$AvailableMetricsGetValuesWithHttpInfo()}}
\item \href{#method-AvailableMetricsApi-clone}{\code{AvailableMetricsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvailableMetricsApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-AvailableMetricsApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AvailableMetricsApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AvailableMetricsApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvailableMetricsApi-AvailableMetricsGetValues"></a>}}
\if{latex}{\out{\hypertarget{method-AvailableMetricsApi-AvailableMetricsGetValues}{}}}
\subsection{Method \code{AvailableMetricsGetValues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AvailableMetricsApi$AvailableMetricsGetValues(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
MetricGroups
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvailableMetricsApi-AvailableMetricsGetValuesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AvailableMetricsApi-AvailableMetricsGetValuesWithHttpInfo}{}}}
\subsection{Method \code{AvailableMetricsGetValuesWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AvailableMetricsApi$AvailableMetricsGetValuesWithHttpInfo(
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (MetricGroups) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AvailableMetricsApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AvailableMetricsApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AvailableMetricsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
