% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/business_combinations_api.R
\docType{class}
\name{BusinessCombinationsApi}
\alias{BusinessCombinationsApi}
\title{BusinessCombinations operations}
\format{
An \code{R6Class} generator object
}
\description{
BusinessCombinationsApi
}
\details{
Calcbench API

No description provided (generated by Openapi Generator https://github.com/calcbenchRtools/calcbenchR-generator)

The version of the OpenAPI document: v1
Generated by: https://calcbenchR-generator.tech
}
\section{Methods}{

\describe{
\strong{ BusinessCombinationsBusinessCombinations } \emph{ The data behind the business combinations page, https://www.calcbench.com/businesscombinations }

\itemize{
\item \emph{ @param } search_parameters \link{BusinessCombinationsParameters}
\item \emph{ @returnType } list( \link{PurchasePriceAllocation} ) \cr

\item status code : 200 | OK

\item return type : array\link{PurchasePriceAllocation}
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  BusinessCombinationsBusinessCombinations  ####################

library(calcbenchR)
var_search_parameters <- BusinessCombinationsParameters$new(123, PeriodParameters$new(123, "unset", 123, "unset", "periodType_example", "useFiscalPeriod_example", DateRange$new("startDate_example", "endDate_example"), "allHistory_example", "updateDate_example", "updatedFrom_example", "asOriginallyReported_example", 123, 123), CompaniesParameters$new(c("companyIdentifiers_example"), "entireUniverse_example", c(123))) # BusinessCombinationsParameters | 

#The data behind the business combinations page, https://www.calcbench.com/businesscombinations
api_instance <- BusinessCombinationsApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$BusinessCombinationsBusinessCombinations(var_search_parametersdata_file = "result.txt")
result <- api_instance$BusinessCombinationsBusinessCombinations(var_search_parameters)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BusinessCombinationsApi-new}{\code{BusinessCombinationsApi$new()}}
\item \href{#method-BusinessCombinationsApi-BusinessCombinationsBusinessCombinations}{\code{BusinessCombinationsApi$BusinessCombinationsBusinessCombinations()}}
\item \href{#method-BusinessCombinationsApi-BusinessCombinationsBusinessCombinationsWithHttpInfo}{\code{BusinessCombinationsApi$BusinessCombinationsBusinessCombinationsWithHttpInfo()}}
\item \href{#method-BusinessCombinationsApi-clone}{\code{BusinessCombinationsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BusinessCombinationsApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsApi-BusinessCombinationsBusinessCombinations"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsApi-BusinessCombinationsBusinessCombinations}{}}}
\subsection{Method \code{BusinessCombinationsBusinessCombinations()}}{
The data behind the business combinations page, https://www.calcbench.com/businesscombinations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsApi$BusinessCombinationsBusinessCombinations(
  search_parameters,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_parameters}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array\link{PurchasePriceAllocation}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsApi-BusinessCombinationsBusinessCombinationsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsApi-BusinessCombinationsBusinessCombinationsWithHttpInfo}{}}}
\subsection{Method \code{BusinessCombinationsBusinessCombinationsWithHttpInfo()}}{
The data behind the business combinations page, https://www.calcbench.com/businesscombinations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsApi$BusinessCombinationsBusinessCombinationsWithHttpInfo(
  search_parameters,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_parameters}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (array\link{PurchasePriceAllocation}) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
