% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/business_combinations_parameters.R
\docType{class}
\name{BusinessCombinationsParameters}
\alias{BusinessCombinationsParameters}
\title{BusinessCombinationsParameters}
\format{
An \code{R6Class} generator object
}
\description{
BusinessCombinationsParameters Class

BusinessCombinationsParameters Class
}
\details{
Create a new BusinessCombinationsParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pageParameters}}{object \link{optional}}

\item{\code{periodParameters}}{\link{PeriodParameters} \link{optional}}

\item{\code{companiesParameters}}{\link{CompaniesParameters} \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BusinessCombinationsParameters-new}{\code{BusinessCombinationsParameters$new()}}
\item \href{#method-BusinessCombinationsParameters-toJSON}{\code{BusinessCombinationsParameters$toJSON()}}
\item \href{#method-BusinessCombinationsParameters-fromJSON}{\code{BusinessCombinationsParameters$fromJSON()}}
\item \href{#method-BusinessCombinationsParameters-toJSONString}{\code{BusinessCombinationsParameters$toJSONString()}}
\item \href{#method-BusinessCombinationsParameters-fromJSONString}{\code{BusinessCombinationsParameters$fromJSONString()}}
\item \href{#method-BusinessCombinationsParameters-validateJSON}{\code{BusinessCombinationsParameters$validateJSON()}}
\item \href{#method-BusinessCombinationsParameters-toString}{\code{BusinessCombinationsParameters$toString()}}
\item \href{#method-BusinessCombinationsParameters-isValid}{\code{BusinessCombinationsParameters$isValid()}}
\item \href{#method-BusinessCombinationsParameters-getInvalidFields}{\code{BusinessCombinationsParameters$getInvalidFields()}}
\item \href{#method-BusinessCombinationsParameters-print}{\code{BusinessCombinationsParameters$print()}}
\item \href{#method-BusinessCombinationsParameters-clone}{\code{BusinessCombinationsParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BusinessCombinationsParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsParameters$new(
  pageParameters = NULL,
  periodParameters = NULL,
  companiesParameters = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pageParameters}}{pageParameters}

\item{\code{periodParameters}}{periodParameters}

\item{\code{companiesParameters}}{companiesParameters}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BusinessCombinationsParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of BusinessCombinationsParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BusinessCombinationsParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BusinessCombinationsParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of BusinessCombinationsParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BusinessCombinationsParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to BusinessCombinationsParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of BusinessCombinationsParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessCombinationsParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessCombinationsParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessCombinationsParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
