% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies_api.R
\docType{class}
\name{CompaniesApi}
\alias{CompaniesApi}
\title{Companies operations}
\format{
An \code{R6Class} generator object
}
\description{
CompaniesApi
}
\details{
Calcbench API

No description provided (generated by Openapi Generator https://github.com/calcbenchRtools/calcbenchR-generator)

The version of the OpenAPI document: v1
Generated by: https://calcbenchR-generator.tech
}
\section{Methods}{

\describe{
\strong{ CompaniesAvailableCompanies } \emph{  }

\itemize{
\item \emph{ @param } search_parameters_user_alert_companies character
\item \emph{ @param } search_parameters_s_ic_codes list( integer )
\item \emph{ @param } search_parameters_index character
\item \emph{ @param } search_parameters_peer_group character
\item \emph{ @param } search_parameters_universe character
\item \emph{ @param } search_parameters_s_ec_only character
\item \emph{ @param } search_parameters_calcbench_entity_ids list( integer )
\item \emph{ @param } search_parameters_name_and_ticker_only character
\item \emph{ @param } search_parameters_include_most_recent_filing_extras character
\item \emph{ @param } search_parameters_company_identifiers list( character )
\item \emph{ @param } search_parameters_n_aics_codes list( integer )
\item \emph{ @returnType } list( \link{APIXBRLCompany} ) \cr

\item status code : 200 | OK

\item return type : array\link{APIXBRLCompany}
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  CompaniesAvailableCompanies  ####################

library(calcbenchR)
var_search_parameters_user_alert_companies <- "search_parameters_user_alert_companies_example" # character |  (Optional)
var_search_parameters_s_ic_codes <- c(123) # array[integer] |  (Optional)
var_search_parameters_index <- "search_parameters_index_example" # character |  (Optional)
var_search_parameters_peer_group <- "search_parameters_peer_group_example" # character |  (Optional)
var_search_parameters_universe <- "search_parameters_universe_example" # character |  (Optional)
var_search_parameters_s_ec_only <- "search_parameters_s_ec_only_example" # character |  (Optional)
var_search_parameters_calcbench_entity_ids <- c(123) # array[integer] |  (Optional)
var_search_parameters_name_and_ticker_only <- "search_parameters_name_and_ticker_only_example" # character |  (Optional)
var_search_parameters_include_most_recent_filing_extras <- "search_parameters_include_most_recent_filing_extras_example" # character |  (Optional)
var_search_parameters_company_identifiers <- c("inner_example") # array[character] |  (Optional)
var_search_parameters_n_aics_codes <- c(123) # array[integer] |  (Optional)

api_instance <- CompaniesApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$CompaniesAvailableCompanies(search_parameters_user_alert_companies = var_search_parameters_user_alert_companies, search_parameters_s_ic_codes = var_search_parameters_s_ic_codes, search_parameters_index = var_search_parameters_index, search_parameters_peer_group = var_search_parameters_peer_group, search_parameters_universe = var_search_parameters_universe, search_parameters_s_ec_only = var_search_parameters_s_ec_only, search_parameters_calcbench_entity_ids = var_search_parameters_calcbench_entity_ids, search_parameters_name_and_ticker_only = var_search_parameters_name_and_ticker_only, search_parameters_include_most_recent_filing_extras = var_search_parameters_include_most_recent_filing_extras, search_parameters_company_identifiers = var_search_parameters_company_identifiers, search_parameters_n_aics_codes = var_search_parameters_n_aics_codesdata_file = "result.txt")
result <- api_instance$CompaniesAvailableCompanies(search_parameters_user_alert_companies = var_search_parameters_user_alert_companies, search_parameters_s_ic_codes = var_search_parameters_s_ic_codes, search_parameters_index = var_search_parameters_index, search_parameters_peer_group = var_search_parameters_peer_group, search_parameters_universe = var_search_parameters_universe, search_parameters_s_ec_only = var_search_parameters_s_ec_only, search_parameters_calcbench_entity_ids = var_search_parameters_calcbench_entity_ids, search_parameters_name_and_ticker_only = var_search_parameters_name_and_ticker_only, search_parameters_include_most_recent_filing_extras = var_search_parameters_include_most_recent_filing_extras, search_parameters_company_identifiers = var_search_parameters_company_identifiers, search_parameters_n_aics_codes = var_search_parameters_n_aics_codes)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CompaniesApi-new}{\code{CompaniesApi$new()}}
\item \href{#method-CompaniesApi-CompaniesAvailableCompanies}{\code{CompaniesApi$CompaniesAvailableCompanies()}}
\item \href{#method-CompaniesApi-CompaniesAvailableCompaniesWithHttpInfo}{\code{CompaniesApi$CompaniesAvailableCompaniesWithHttpInfo()}}
\item \href{#method-CompaniesApi-clone}{\code{CompaniesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CompaniesApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesApi-CompaniesAvailableCompanies"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesApi-CompaniesAvailableCompanies}{}}}
\subsection{Method \code{CompaniesAvailableCompanies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesApi$CompaniesAvailableCompanies(
  search_parameters_user_alert_companies = NULL,
  search_parameters_s_ic_codes = NULL,
  search_parameters_index = NULL,
  search_parameters_peer_group = NULL,
  search_parameters_universe = NULL,
  search_parameters_s_ec_only = NULL,
  search_parameters_calcbench_entity_ids = NULL,
  search_parameters_name_and_ticker_only = NULL,
  search_parameters_include_most_recent_filing_extras = NULL,
  search_parameters_company_identifiers = NULL,
  search_parameters_n_aics_codes = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_parameters_user_alert_companies}}{(optional) No description}

\item{\code{search_parameters_s_ic_codes}}{(optional) No description}

\item{\code{search_parameters_index}}{(optional) No description}

\item{\code{search_parameters_peer_group}}{(optional) No description}

\item{\code{search_parameters_universe}}{(optional) No description}

\item{\code{search_parameters_s_ec_only}}{(optional) No description}

\item{\code{search_parameters_calcbench_entity_ids}}{(optional) No description}

\item{\code{search_parameters_name_and_ticker_only}}{(optional) No description}

\item{\code{search_parameters_include_most_recent_filing_extras}}{(optional) No description}

\item{\code{search_parameters_company_identifiers}}{(optional) No description}

\item{\code{search_parameters_n_aics_codes}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array\link{APIXBRLCompany}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesApi-CompaniesAvailableCompaniesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesApi-CompaniesAvailableCompaniesWithHttpInfo}{}}}
\subsection{Method \code{CompaniesAvailableCompaniesWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesApi$CompaniesAvailableCompaniesWithHttpInfo(
  search_parameters_user_alert_companies = NULL,
  search_parameters_s_ic_codes = NULL,
  search_parameters_index = NULL,
  search_parameters_peer_group = NULL,
  search_parameters_universe = NULL,
  search_parameters_s_ec_only = NULL,
  search_parameters_calcbench_entity_ids = NULL,
  search_parameters_name_and_ticker_only = NULL,
  search_parameters_include_most_recent_filing_extras = NULL,
  search_parameters_company_identifiers = NULL,
  search_parameters_n_aics_codes = NULL,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_parameters_user_alert_companies}}{(optional) No description}

\item{\code{search_parameters_s_ic_codes}}{(optional) No description}

\item{\code{search_parameters_index}}{(optional) No description}

\item{\code{search_parameters_peer_group}}{(optional) No description}

\item{\code{search_parameters_universe}}{(optional) No description}

\item{\code{search_parameters_s_ec_only}}{(optional) No description}

\item{\code{search_parameters_calcbench_entity_ids}}{(optional) No description}

\item{\code{search_parameters_name_and_ticker_only}}{(optional) No description}

\item{\code{search_parameters_include_most_recent_filing_extras}}{(optional) No description}

\item{\code{search_parameters_company_identifiers}}{(optional) No description}

\item{\code{search_parameters_n_aics_codes}}{(optional) No description}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (array\link{APIXBRLCompany}) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
