% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies_parameters.R
\docType{class}
\name{CompaniesParameters}
\alias{CompaniesParameters}
\title{CompaniesParameters}
\format{
An \code{R6Class} generator object
}
\description{
CompaniesParameters Class

CompaniesParameters Class
}
\details{
Create a new CompaniesParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{companyIdentifiers}}{list(character) \link{optional}}

\item{\code{entireUniverse}}{character \link{optional}}

\item{\code{calcbenchEntityIDs}}{list(integer) \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CompaniesParameters-new}{\code{CompaniesParameters$new()}}
\item \href{#method-CompaniesParameters-toJSON}{\code{CompaniesParameters$toJSON()}}
\item \href{#method-CompaniesParameters-fromJSON}{\code{CompaniesParameters$fromJSON()}}
\item \href{#method-CompaniesParameters-toJSONString}{\code{CompaniesParameters$toJSONString()}}
\item \href{#method-CompaniesParameters-fromJSONString}{\code{CompaniesParameters$fromJSONString()}}
\item \href{#method-CompaniesParameters-validateJSON}{\code{CompaniesParameters$validateJSON()}}
\item \href{#method-CompaniesParameters-toString}{\code{CompaniesParameters$toString()}}
\item \href{#method-CompaniesParameters-isValid}{\code{CompaniesParameters$isValid()}}
\item \href{#method-CompaniesParameters-getInvalidFields}{\code{CompaniesParameters$getInvalidFields()}}
\item \href{#method-CompaniesParameters-print}{\code{CompaniesParameters$print()}}
\item \href{#method-CompaniesParameters-clone}{\code{CompaniesParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CompaniesParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesParameters$new(
  companyIdentifiers = NULL,
  entireUniverse = NULL,
  calcbenchEntityIDs = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{companyIdentifiers}}{companyIdentifiers}

\item{\code{entireUniverse}}{entireUniverse}

\item{\code{calcbenchEntityIDs}}{calcbenchEntityIDs}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CompaniesParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of CompaniesParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CompaniesParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CompaniesParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of CompaniesParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CompaniesParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to CompaniesParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of CompaniesParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompaniesParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CompaniesParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompaniesParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
