% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_search_parameters.R
\docType{class}
\name{CompanySearchParameters}
\alias{CompanySearchParameters}
\title{CompanySearchParameters}
\format{
An \code{R6Class} generator object
}
\description{
CompanySearchParameters Class

CompanySearchParameters Class
}
\details{
Create a new CompanySearchParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{userAlertCompanies}}{character \link{optional}}

\item{\code{SICCodes}}{list(integer) \link{optional}}

\item{\code{index}}{character \link{optional}}

\item{\code{peerGroup}}{character \link{optional}}

\item{\code{universe}}{character \link{optional}}

\item{\code{SECOnly}}{character \link{optional}}

\item{\code{calcbenchEntityIds}}{list(integer) \link{optional}}

\item{\code{nameAndTickerOnly}}{character \link{optional}}

\item{\code{includeMostRecentFilingExtras}}{character \link{optional}}

\item{\code{companyIdentifiers}}{list(character) \link{optional}}

\item{\code{NAICSCodes}}{list(integer) \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CompanySearchParameters-new}{\code{CompanySearchParameters$new()}}
\item \href{#method-CompanySearchParameters-toJSON}{\code{CompanySearchParameters$toJSON()}}
\item \href{#method-CompanySearchParameters-fromJSON}{\code{CompanySearchParameters$fromJSON()}}
\item \href{#method-CompanySearchParameters-toJSONString}{\code{CompanySearchParameters$toJSONString()}}
\item \href{#method-CompanySearchParameters-fromJSONString}{\code{CompanySearchParameters$fromJSONString()}}
\item \href{#method-CompanySearchParameters-validateJSON}{\code{CompanySearchParameters$validateJSON()}}
\item \href{#method-CompanySearchParameters-toString}{\code{CompanySearchParameters$toString()}}
\item \href{#method-CompanySearchParameters-isValid}{\code{CompanySearchParameters$isValid()}}
\item \href{#method-CompanySearchParameters-getInvalidFields}{\code{CompanySearchParameters$getInvalidFields()}}
\item \href{#method-CompanySearchParameters-print}{\code{CompanySearchParameters$print()}}
\item \href{#method-CompanySearchParameters-clone}{\code{CompanySearchParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanySearchParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-CompanySearchParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CompanySearchParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanySearchParameters$new(
  userAlertCompanies = NULL,
  SICCodes = NULL,
  index = NULL,
  peerGroup = NULL,
  universe = NULL,
  SECOnly = NULL,
  calcbenchEntityIds = NULL,
  nameAndTickerOnly = NULL,
  includeMostRecentFilingExtras = NULL,
  companyIdentifiers = NULL,
  NAICSCodes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userAlertCompanies}}{userAlertCompanies}

\item{\code{SICCodes}}{SICCodes}

\item{\code{index}}{index}

\item{\code{peerGroup}}{peerGroup}

\item{\code{universe}}{universe}

\item{\code{SECOnly}}{SECOnly}

\item{\code{calcbenchEntityIds}}{calcbenchEntityIds}

\item{\code{nameAndTickerOnly}}{nameAndTickerOnly}

\item{\code{includeMostRecentFilingExtras}}{includeMostRecentFilingExtras}

\item{\code{companyIdentifiers}}{companyIdentifiers}

\item{\code{NAICSCodes}}{NAICSCodes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanySearchParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CompanySearchParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanySearchParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CompanySearchParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanySearchParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CompanySearchParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of CompanySearchParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanySearchParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CompanySearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanySearchParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CompanySearchParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanySearchParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CompanySearchParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanySearchParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CompanySearchParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of CompanySearchParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanySearchParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CompanySearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanySearchParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CompanySearchParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to CompanySearchParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanySearchParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanySearchParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-CompanySearchParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanySearchParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of CompanySearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanySearchParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-CompanySearchParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanySearchParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanySearchParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-CompanySearchParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanySearchParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanySearchParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-CompanySearchParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanySearchParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanySearchParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CompanySearchParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanySearchParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
