% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_data_api.R
\docType{class}
\name{DimensionalDataApi}
\alias{DimensionalDataApi}
\title{DimensionalData operations}
\format{
An \code{R6Class} generator object
}
\description{
DimensionalDataApi
}
\details{
Calcbench API

No description provided (generated by Openapi Generator https://github.com/calcbenchRtools/calcbenchR-generator)

The version of the OpenAPI document: v1
Generated by: https://calcbenchR-generator.tech
}
\section{Methods}{

\describe{
\strong{ DimensionalDataDimensionalData } \emph{ The data behind https://www.calcbench.com/segments }

\itemize{
\item \emph{ @param } params \link{DimensionalSearchParameters}
\item \emph{ @returnType } list( \link{DimensionalDataPoint} ) \cr

\item status code : 200 | OK

\item return type : array\link{DimensionalDataPoint}
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  DimensionalDataDimensionalData  ####################

library(calcbenchR)
var_params <- DimensionalSearchParameters$new(DimensionalQueryParameters$new("searchTerms_example", c("metrics_example"), 123, 123, "queryID_example", "AsOriginallyReported_example", "dimensionName_example", "dataSet_example"), PeriodParameters$new(123, "unset", 123, "unset", "periodType_example", "useFiscalPeriod_example", DateRange$new("startDate_example", "endDate_example"), "allHistory_example", "updateDate_example", "updatedFrom_example", "asOriginallyReported_example", 123, 123), CompaniesParameters$new(c("companyIdentifiers_example"), "entireUniverse_example", c(123))) # DimensionalSearchParameters | 

#The data behind https://www.calcbench.com/segments
api_instance <- DimensionalDataApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$DimensionalDataDimensionalData(var_paramsdata_file = "result.txt")
result <- api_instance$DimensionalDataDimensionalData(var_params)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DimensionalDataApi-new}{\code{DimensionalDataApi$new()}}
\item \href{#method-DimensionalDataApi-DimensionalDataDimensionalData}{\code{DimensionalDataApi$DimensionalDataDimensionalData()}}
\item \href{#method-DimensionalDataApi-DimensionalDataDimensionalDataWithHttpInfo}{\code{DimensionalDataApi$DimensionalDataDimensionalDataWithHttpInfo()}}
\item \href{#method-DimensionalDataApi-clone}{\code{DimensionalDataApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalDataApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalDataApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DimensionalDataApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalDataApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalDataApi-DimensionalDataDimensionalData"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalDataApi-DimensionalDataDimensionalData}{}}}
\subsection{Method \code{DimensionalDataDimensionalData()}}{
The data behind https://www.calcbench.com/segments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalDataApi$DimensionalDataDimensionalData(
  params,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array\link{DimensionalDataPoint}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalDataApi-DimensionalDataDimensionalDataWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalDataApi-DimensionalDataDimensionalDataWithHttpInfo}{}}}
\subsection{Method \code{DimensionalDataDimensionalDataWithHttpInfo()}}{
The data behind https://www.calcbench.com/segments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalDataApi$DimensionalDataDimensionalDataWithHttpInfo(
  params,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (array\link{DimensionalDataPoint}) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalDataApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalDataApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalDataApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
