% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_query_parameters.R
\docType{class}
\name{DimensionalQueryParameters}
\alias{DimensionalQueryParameters}
\title{DimensionalQueryParameters}
\format{
An \code{R6Class} generator object
}
\description{
DimensionalQueryParameters Class

DimensionalQueryParameters Class
}
\details{
Create a new DimensionalQueryParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{searchTerms}}{character \link{optional}}

\item{\code{metrics}}{list(character) \link{optional}}

\item{\code{filterValue}}{numeric \link{optional}}

\item{\code{filterOperator}}{integer \link{optional}}

\item{\code{queryID}}{character \link{optional}}

\item{\code{AsOriginallyReported}}{Show the first reported, rather than revised, values character \link{optional}}

\item{\code{dimensionName}}{character \link{optional}}

\item{\code{dataSet}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DimensionalQueryParameters-new}{\code{DimensionalQueryParameters$new()}}
\item \href{#method-DimensionalQueryParameters-toJSON}{\code{DimensionalQueryParameters$toJSON()}}
\item \href{#method-DimensionalQueryParameters-fromJSON}{\code{DimensionalQueryParameters$fromJSON()}}
\item \href{#method-DimensionalQueryParameters-toJSONString}{\code{DimensionalQueryParameters$toJSONString()}}
\item \href{#method-DimensionalQueryParameters-fromJSONString}{\code{DimensionalQueryParameters$fromJSONString()}}
\item \href{#method-DimensionalQueryParameters-validateJSON}{\code{DimensionalQueryParameters$validateJSON()}}
\item \href{#method-DimensionalQueryParameters-toString}{\code{DimensionalQueryParameters$toString()}}
\item \href{#method-DimensionalQueryParameters-isValid}{\code{DimensionalQueryParameters$isValid()}}
\item \href{#method-DimensionalQueryParameters-getInvalidFields}{\code{DimensionalQueryParameters$getInvalidFields()}}
\item \href{#method-DimensionalQueryParameters-print}{\code{DimensionalQueryParameters$print()}}
\item \href{#method-DimensionalQueryParameters-clone}{\code{DimensionalQueryParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalQueryParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalQueryParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DimensionalQueryParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalQueryParameters$new(
  searchTerms = NULL,
  metrics = NULL,
  filterValue = NULL,
  filterOperator = NULL,
  queryID = NULL,
  AsOriginallyReported = NULL,
  dimensionName = NULL,
  dataSet = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchTerms}}{searchTerms}

\item{\code{metrics}}{metrics}

\item{\code{filterValue}}{filterValue}

\item{\code{filterOperator}}{filterOperator}

\item{\code{queryID}}{queryID}

\item{\code{AsOriginallyReported}}{Show the first reported, rather than revised, values}

\item{\code{dimensionName}}{dimensionName}

\item{\code{dataSet}}{dataSet}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalQueryParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalQueryParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalQueryParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DimensionalQueryParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalQueryParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalQueryParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DimensionalQueryParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalQueryParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DimensionalQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalQueryParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalQueryParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalQueryParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DimensionalQueryParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalQueryParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalQueryParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DimensionalQueryParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalQueryParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DimensionalQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalQueryParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalQueryParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to DimensionalQueryParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalQueryParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalQueryParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalQueryParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalQueryParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of DimensionalQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalQueryParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalQueryParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalQueryParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalQueryParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalQueryParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalQueryParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalQueryParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalQueryParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalQueryParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalQueryParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalQueryParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalQueryParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
