% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_search_parameters.R
\docType{class}
\name{DimensionalSearchParameters}
\alias{DimensionalSearchParameters}
\title{DimensionalSearchParameters}
\format{
An \code{R6Class} generator object
}
\description{
The arguments for the Segment API end-point

DimensionalSearchParameters Class
}
\details{
Create a new DimensionalSearchParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pageParameters}}{\link{DimensionalQueryParameters} \link{optional}}

\item{\code{periodParameters}}{\link{PeriodParameters} \link{optional}}

\item{\code{companiesParameters}}{\link{CompaniesParameters} \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DimensionalSearchParameters-new}{\code{DimensionalSearchParameters$new()}}
\item \href{#method-DimensionalSearchParameters-toJSON}{\code{DimensionalSearchParameters$toJSON()}}
\item \href{#method-DimensionalSearchParameters-fromJSON}{\code{DimensionalSearchParameters$fromJSON()}}
\item \href{#method-DimensionalSearchParameters-toJSONString}{\code{DimensionalSearchParameters$toJSONString()}}
\item \href{#method-DimensionalSearchParameters-fromJSONString}{\code{DimensionalSearchParameters$fromJSONString()}}
\item \href{#method-DimensionalSearchParameters-validateJSON}{\code{DimensionalSearchParameters$validateJSON()}}
\item \href{#method-DimensionalSearchParameters-toString}{\code{DimensionalSearchParameters$toString()}}
\item \href{#method-DimensionalSearchParameters-isValid}{\code{DimensionalSearchParameters$isValid()}}
\item \href{#method-DimensionalSearchParameters-getInvalidFields}{\code{DimensionalSearchParameters$getInvalidFields()}}
\item \href{#method-DimensionalSearchParameters-print}{\code{DimensionalSearchParameters$print()}}
\item \href{#method-DimensionalSearchParameters-clone}{\code{DimensionalSearchParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalSearchParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalSearchParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DimensionalSearchParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalSearchParameters$new(
  pageParameters = NULL,
  periodParameters = NULL,
  companiesParameters = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pageParameters}}{pageParameters}

\item{\code{periodParameters}}{periodParameters}

\item{\code{companiesParameters}}{companiesParameters}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalSearchParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalSearchParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalSearchParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DimensionalSearchParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalSearchParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalSearchParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DimensionalSearchParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalSearchParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DimensionalSearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalSearchParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalSearchParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalSearchParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DimensionalSearchParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalSearchParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalSearchParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DimensionalSearchParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalSearchParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DimensionalSearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalSearchParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalSearchParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to DimensionalSearchParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalSearchParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalSearchParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalSearchParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalSearchParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of DimensionalSearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalSearchParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalSearchParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalSearchParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalSearchParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalSearchParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalSearchParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalSearchParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalSearchParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalSearchParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionalSearchParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionalSearchParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionalSearchParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
