% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_api.R
\docType{class}
\name{DisclosureApi}
\alias{DisclosureApi}
\title{Disclosure operations}
\format{
An \code{R6Class} generator object
}
\description{
DisclosureApi
}
\details{
Calcbench API

No description provided (generated by Openapi Generator https://github.com/calcbenchRtools/calcbenchR-generator)

The version of the OpenAPI document: v1
Generated by: https://calcbenchR-generator.tech
}
\section{Methods}{

\describe{
\strong{ DisclosureFootnoteSearch } \emph{ Document Search }
The search behind the <a href=\"/disclosures\">disclosures\if{html}{\out{</a>}} page.  To get all results keep calling with \if{html}{\out{<code>}}nextGroupStartOffset\if{html}{\out{</code>}} from the previous response, until \if{html}{\out{<code>}}moreResults\if{html}{\out{</code>}} is false.  Retrieving the document contents varies depending on the type of document returned.

\itemize{
\item \emph{ @param } search_parameters \link{FootnoteSearchParameters}
\item \emph{ @returnType } \link{FootnoteQueryResponse} \cr

\item status code : 200 | OK

\item return type : FootnoteQueryResponse
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  DisclosureFootnoteSearch  ####################

library(calcbenchR)
var_search_parameters <- FootnoteSearchParameters$new(FootnoteQueryParameters$new("fullTextQuery_example", c(123), "footnoteTag_example", 123, 123, 123, "allFootnotes_example", "includeRelatedItems_example", "fullTextSearchAllTimePeriods_example", "fullTextSearchRestrictToMetricLabels_example", "normalizedFaceMetric_example", "normalizedFootnoteMetric_example", "normalizedBreakoutMetric_example", "ByDocumentID_example", "XBRLNumericTag_example", 123, "blobID_example", 123, "nonXBRLExtractTag_example", 123, "userSubmittedURL_example", "subDivide_example", "publicDisclosureID_example", 123, "disclosureName_example", c("disclosureNames_example"), 123, "none", 123, "AllTextBlocks_example", "AllMDASections_example", 123), FootnotePeriodParameters$new("unset", 123, 123, "unset", "periodType_example", "useFiscalPeriod_example", DateRange$new("startDate_example", "endDate_example"), "allHistory_example", "updateDate_example", "updatedFrom_example", "asOriginallyReported_example", 123, 123), CompaniesParameters$new(c("companyIdentifiers_example"), "entireUniverse_example", c(123))) # FootnoteSearchParameters | 

#Document Search
api_instance <- DisclosureApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$DisclosureFootnoteSearch(var_search_parametersdata_file = "result.txt")
result <- api_instance$DisclosureFootnoteSearch(var_search_parameters)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DisclosureApi-new}{\code{DisclosureApi$new()}}
\item \href{#method-DisclosureApi-DisclosureFootnoteSearch}{\code{DisclosureApi$DisclosureFootnoteSearch()}}
\item \href{#method-DisclosureApi-DisclosureFootnoteSearchWithHttpInfo}{\code{DisclosureApi$DisclosureFootnoteSearchWithHttpInfo()}}
\item \href{#method-DisclosureApi-clone}{\code{DisclosureApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DisclosureApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureApi-DisclosureFootnoteSearch"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureApi-DisclosureFootnoteSearch}{}}}
\subsection{Method \code{DisclosureFootnoteSearch()}}{
Document Search
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureApi$DisclosureFootnoteSearch(
  search_parameters,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_parameters}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
FootnoteQueryResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureApi-DisclosureFootnoteSearchWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureApi-DisclosureFootnoteSearchWithHttpInfo}{}}}
\subsection{Method \code{DisclosureFootnoteSearchWithHttpInfo()}}{
Document Search
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureApi$DisclosureFootnoteSearchWithHttpInfo(
  search_parameters,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_parameters}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (FootnoteQueryResponse) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
