% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_sub_section.R
\docType{class}
\name{DisclosureSubSection}
\alias{DisclosureSubSection}
\title{DisclosureSubSection}
\format{
An \code{R6Class} generator object
}
\description{
DisclosureSubSection Class

DisclosureSubSection Class
}
\details{
Create a new DisclosureSubSection
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{header}}{character \link{optional}}

\item{\code{body}}{character \link{optional}}

\item{\code{type}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DisclosureSubSection-new}{\code{DisclosureSubSection$new()}}
\item \href{#method-DisclosureSubSection-toJSON}{\code{DisclosureSubSection$toJSON()}}
\item \href{#method-DisclosureSubSection-fromJSON}{\code{DisclosureSubSection$fromJSON()}}
\item \href{#method-DisclosureSubSection-toJSONString}{\code{DisclosureSubSection$toJSONString()}}
\item \href{#method-DisclosureSubSection-fromJSONString}{\code{DisclosureSubSection$fromJSONString()}}
\item \href{#method-DisclosureSubSection-validateJSON}{\code{DisclosureSubSection$validateJSON()}}
\item \href{#method-DisclosureSubSection-toString}{\code{DisclosureSubSection$toString()}}
\item \href{#method-DisclosureSubSection-isValid}{\code{DisclosureSubSection$isValid()}}
\item \href{#method-DisclosureSubSection-getInvalidFields}{\code{DisclosureSubSection$getInvalidFields()}}
\item \href{#method-DisclosureSubSection-print}{\code{DisclosureSubSection$print()}}
\item \href{#method-DisclosureSubSection-clone}{\code{DisclosureSubSection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureSubSection-new"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureSubSection-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DisclosureSubSection class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureSubSection$new(header = NULL, body = NULL, type = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{header}}{header}

\item{\code{body}}{body}

\item{\code{type}}{type}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureSubSection-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureSubSection-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureSubSection$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DisclosureSubSection in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureSubSection-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureSubSection-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DisclosureSubSection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureSubSection$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DisclosureSubSection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureSubSection-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureSubSection-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureSubSection$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DisclosureSubSection in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureSubSection-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureSubSection-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DisclosureSubSection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureSubSection$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DisclosureSubSection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureSubSection-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureSubSection-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to DisclosureSubSection and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureSubSection$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureSubSection-toString"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureSubSection-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureSubSection$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of DisclosureSubSection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureSubSection-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureSubSection-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureSubSection$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureSubSection-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureSubSection-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureSubSection$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureSubSection-print"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureSubSection-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureSubSection$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisclosureSubSection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DisclosureSubSection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisclosureSubSection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
