% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filing.R
\docType{class}
\name{Filing}
\alias{Filing}
\title{Filing}
\format{
An \code{R6Class} generator object
}
\description{
Filing Class

Filing Class
}
\details{
Create a new Filing
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{is_xbrl}}{character \link{optional}}

\item{\code{is_wire}}{character \link{optional}}

\item{\code{calcbench_id}}{integer \link{optional}}

\item{\code{filing_id}}{integer \link{optional}}

\item{\code{sec_accession_id}}{character \link{optional}}

\item{\code{sec_html_url}}{character \link{optional}}

\item{\code{document_type}}{character \link{optional}}

\item{\code{filing_type}}{character \link{optional}}

\item{\code{filing_date}}{character \link{optional}}

\item{\code{fiscal_period}}{integer \link{optional}}

\item{\code{fiscal_year}}{integer \link{optional}}

\item{\code{calcbench_accepted}}{character \link{optional}}

\item{\code{calcbench_finished_load}}{character \link{optional}}

\item{\code{entity_id}}{integer \link{optional}}

\item{\code{ticker}}{character \link{optional}}

\item{\code{entity_name}}{character \link{optional}}

\item{\code{CIK}}{character \link{optional}}

\item{\code{period_index}}{integer \link{optional}}

\item{\code{associated_proxy_SEC_URL}}{character \link{optional}}

\item{\code{associated_earnings_press_release_SEC_URL}}{character \link{optional}}

\item{\code{period_end_date}}{character \link{optional}}

\item{\code{percentage_revenue_change}}{numeric \link{optional}}

\item{\code{this_period_revenue}}{numeric \link{optional}}

\item{\code{link1}}{character \link{optional}}

\item{\code{link2}}{character \link{optional}}

\item{\code{link3}}{character \link{optional}}

\item{\code{calendar_year}}{integer \link{optional}}

\item{\code{calendar_period}}{integer \link{optional}}

\item{\code{standardized_XBRL}}{character \link{optional}}

\item{\code{has_standardized_data}}{character \link{optional}}

\item{\code{item_types}}{list(character) \link{optional}}

\item{\code{standardized_data_compressed}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Filing-new}{\code{Filing$new()}}
\item \href{#method-Filing-toJSON}{\code{Filing$toJSON()}}
\item \href{#method-Filing-fromJSON}{\code{Filing$fromJSON()}}
\item \href{#method-Filing-toJSONString}{\code{Filing$toJSONString()}}
\item \href{#method-Filing-fromJSONString}{\code{Filing$fromJSONString()}}
\item \href{#method-Filing-validateJSON}{\code{Filing$validateJSON()}}
\item \href{#method-Filing-toString}{\code{Filing$toString()}}
\item \href{#method-Filing-isValid}{\code{Filing$isValid()}}
\item \href{#method-Filing-getInvalidFields}{\code{Filing$getInvalidFields()}}
\item \href{#method-Filing-print}{\code{Filing$print()}}
\item \href{#method-Filing-clone}{\code{Filing$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filing-new"></a>}}
\if{latex}{\out{\hypertarget{method-Filing-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Filing class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filing$new(
  is_xbrl = NULL,
  is_wire = NULL,
  calcbench_id = NULL,
  filing_id = NULL,
  sec_accession_id = NULL,
  sec_html_url = NULL,
  document_type = NULL,
  filing_type = NULL,
  filing_date = NULL,
  fiscal_period = NULL,
  fiscal_year = NULL,
  calcbench_accepted = NULL,
  calcbench_finished_load = NULL,
  entity_id = NULL,
  ticker = NULL,
  entity_name = NULL,
  CIK = NULL,
  period_index = NULL,
  associated_proxy_SEC_URL = NULL,
  associated_earnings_press_release_SEC_URL = NULL,
  period_end_date = NULL,
  percentage_revenue_change = NULL,
  this_period_revenue = NULL,
  link1 = NULL,
  link2 = NULL,
  link3 = NULL,
  calendar_year = NULL,
  calendar_period = NULL,
  standardized_XBRL = NULL,
  has_standardized_data = NULL,
  item_types = NULL,
  standardized_data_compressed = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{is_xbrl}}{is_xbrl}

\item{\code{is_wire}}{is_wire}

\item{\code{calcbench_id}}{calcbench_id}

\item{\code{filing_id}}{filing_id}

\item{\code{sec_accession_id}}{sec_accession_id}

\item{\code{sec_html_url}}{sec_html_url}

\item{\code{document_type}}{document_type}

\item{\code{filing_type}}{filing_type}

\item{\code{filing_date}}{filing_date}

\item{\code{fiscal_period}}{fiscal_period}

\item{\code{fiscal_year}}{fiscal_year}

\item{\code{calcbench_accepted}}{calcbench_accepted}

\item{\code{calcbench_finished_load}}{calcbench_finished_load}

\item{\code{entity_id}}{entity_id}

\item{\code{ticker}}{ticker}

\item{\code{entity_name}}{entity_name}

\item{\code{CIK}}{CIK}

\item{\code{period_index}}{period_index}

\item{\code{associated_proxy_SEC_URL}}{associated_proxy_SEC_URL}

\item{\code{associated_earnings_press_release_SEC_URL}}{associated_earnings_press_release_SEC_URL}

\item{\code{period_end_date}}{period_end_date}

\item{\code{percentage_revenue_change}}{percentage_revenue_change}

\item{\code{this_period_revenue}}{this_period_revenue}

\item{\code{link1}}{link1}

\item{\code{link2}}{link2}

\item{\code{link3}}{link3}

\item{\code{calendar_year}}{calendar_year}

\item{\code{calendar_period}}{calendar_period}

\item{\code{standardized_XBRL}}{standardized_XBRL}

\item{\code{has_standardized_data}}{has_standardized_data}

\item{\code{item_types}}{item_types}

\item{\code{standardized_data_compressed}}{standardized_data_compressed}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filing-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Filing-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filing$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Filing in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filing-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Filing-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Filing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filing$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Filing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filing-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Filing-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filing$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Filing in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filing-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Filing-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Filing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filing$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Filing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filing-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Filing-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Filing and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filing$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filing-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Filing-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filing$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Filing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filing-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Filing-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filing$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filing-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Filing-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filing$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filing-print"></a>}}
\if{latex}{\out{\hypertarget{method-Filing-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filing$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filing-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Filing-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filing$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
