% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filings_api.R
\docType{class}
\name{FilingsApi}
\alias{FilingsApi}
\title{Filings operations}
\format{
An \code{R6Class} generator object
}
\description{
FilingsApi
}
\details{
Calcbench API

No description provided (generated by Openapi Generator https://github.com/calcbenchRtools/calcbenchR-generator)

The version of the OpenAPI document: v1
Generated by: https://calcbenchR-generator.tech
}
\section{Methods}{

\describe{
\strong{ FilingsFilingsV2 } \emph{  }

\itemize{
\item \emph{ @param } filings_params \link{FilingsParameters}
\item \emph{ @returnType } list( \link{Filing} ) \cr

\item status code : 200 | OK

\item return type : array\link{Filing}
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  FilingsFilingsV2  ####################

library(calcbenchR)
var_filings_params <- FilingsParameters$new(CompaniesParameters$new(c("companyIdentifiers_example"), "entireUniverse_example", c(123)), FilingsQueryParameters$new("IncludePerformance_example", "IncludePressReleasesAndProxies_example", "IncludeNonXBRL_example", c("BusinessWirePR_filedAfterAn8K"), 123, c("metrics_example"), "IncludeXBRL_example"), PeriodParameters$new(123, "unset", 123, "unset", "periodType_example", "useFiscalPeriod_example", DateRange$new("startDate_example", "endDate_example"), "allHistory_example", "updateDate_example", "updatedFrom_example", "asOriginallyReported_example", 123, 123)) # FilingsParameters | 

api_instance <- FilingsApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$FilingsFilingsV2(var_filings_paramsdata_file = "result.txt")
result <- api_instance$FilingsFilingsV2(var_filings_params)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilingsApi-new}{\code{FilingsApi$new()}}
\item \href{#method-FilingsApi-FilingsFilingsV2}{\code{FilingsApi$FilingsFilingsV2()}}
\item \href{#method-FilingsApi-FilingsFilingsV2WithHttpInfo}{\code{FilingsApi$FilingsFilingsV2WithHttpInfo()}}
\item \href{#method-FilingsApi-clone}{\code{FilingsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FilingsApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsApi-FilingsFilingsV2"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsApi-FilingsFilingsV2}{}}}
\subsection{Method \code{FilingsFilingsV2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsApi$FilingsFilingsV2(filings_params, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filings_params}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array\link{Filing}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsApi-FilingsFilingsV2WithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsApi-FilingsFilingsV2WithHttpInfo}{}}}
\subsection{Method \code{FilingsFilingsV2WithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsApi$FilingsFilingsV2WithHttpInfo(filings_params, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filings_params}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (array\link{Filing}) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
