% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filings_parameters.R
\docType{class}
\name{FilingsParameters}
\alias{FilingsParameters}
\title{FilingsParameters}
\format{
An \code{R6Class} generator object
}
\description{
FilingsParameters Class

FilingsParameters Class
}
\details{
Create a new FilingsParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{companiesParameters}}{\link{CompaniesParameters} \link{optional}}

\item{\code{pageParameters}}{\link{FilingsQueryParameters} \link{optional}}

\item{\code{periodParameters}}{\link{PeriodParameters} \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilingsParameters-new}{\code{FilingsParameters$new()}}
\item \href{#method-FilingsParameters-toJSON}{\code{FilingsParameters$toJSON()}}
\item \href{#method-FilingsParameters-fromJSON}{\code{FilingsParameters$fromJSON()}}
\item \href{#method-FilingsParameters-toJSONString}{\code{FilingsParameters$toJSONString()}}
\item \href{#method-FilingsParameters-fromJSONString}{\code{FilingsParameters$fromJSONString()}}
\item \href{#method-FilingsParameters-validateJSON}{\code{FilingsParameters$validateJSON()}}
\item \href{#method-FilingsParameters-toString}{\code{FilingsParameters$toString()}}
\item \href{#method-FilingsParameters-isValid}{\code{FilingsParameters$isValid()}}
\item \href{#method-FilingsParameters-getInvalidFields}{\code{FilingsParameters$getInvalidFields()}}
\item \href{#method-FilingsParameters-print}{\code{FilingsParameters$print()}}
\item \href{#method-FilingsParameters-clone}{\code{FilingsParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FilingsParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsParameters$new(
  companiesParameters = NULL,
  pageParameters = NULL,
  periodParameters = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{companiesParameters}}{companiesParameters}

\item{\code{pageParameters}}{pageParameters}

\item{\code{periodParameters}}{periodParameters}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FilingsParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FilingsParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FilingsParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FilingsParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FilingsParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FilingsParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FilingsParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FilingsParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
