% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filings_query_parameters.R
\docType{class}
\name{FilingsQueryParameters}
\alias{FilingsQueryParameters}
\title{FilingsQueryParameters}
\format{
An \code{R6Class} generator object
}
\description{
FilingsQueryParameters Class

FilingsQueryParameters Class
}
\details{
Create a new FilingsQueryParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{IncludePerformance}}{character \link{optional}}

\item{\code{IncludePressReleasesAndProxies}}{character \link{optional}}

\item{\code{IncludeNonXBRL}}{character \link{optional}}

\item{\code{FilingTypes}}{list(character) \link{optional}}

\item{\code{limit}}{integer \link{optional}}

\item{\code{metrics}}{list(character) \link{optional}}

\item{\code{IncludeXBRL}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilingsQueryParameters-new}{\code{FilingsQueryParameters$new()}}
\item \href{#method-FilingsQueryParameters-toJSON}{\code{FilingsQueryParameters$toJSON()}}
\item \href{#method-FilingsQueryParameters-fromJSON}{\code{FilingsQueryParameters$fromJSON()}}
\item \href{#method-FilingsQueryParameters-toJSONString}{\code{FilingsQueryParameters$toJSONString()}}
\item \href{#method-FilingsQueryParameters-fromJSONString}{\code{FilingsQueryParameters$fromJSONString()}}
\item \href{#method-FilingsQueryParameters-validateJSON}{\code{FilingsQueryParameters$validateJSON()}}
\item \href{#method-FilingsQueryParameters-toString}{\code{FilingsQueryParameters$toString()}}
\item \href{#method-FilingsQueryParameters-isValid}{\code{FilingsQueryParameters$isValid()}}
\item \href{#method-FilingsQueryParameters-getInvalidFields}{\code{FilingsQueryParameters$getInvalidFields()}}
\item \href{#method-FilingsQueryParameters-print}{\code{FilingsQueryParameters$print()}}
\item \href{#method-FilingsQueryParameters-clone}{\code{FilingsQueryParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsQueryParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsQueryParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FilingsQueryParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsQueryParameters$new(
  IncludePerformance = NULL,
  IncludePressReleasesAndProxies = NULL,
  IncludeNonXBRL = NULL,
  FilingTypes = NULL,
  limit = NULL,
  metrics = NULL,
  IncludeXBRL = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{IncludePerformance}}{IncludePerformance}

\item{\code{IncludePressReleasesAndProxies}}{IncludePressReleasesAndProxies}

\item{\code{IncludeNonXBRL}}{IncludeNonXBRL}

\item{\code{FilingTypes}}{FilingTypes}

\item{\code{limit}}{limit}

\item{\code{metrics}}{metrics}

\item{\code{IncludeXBRL}}{IncludeXBRL}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsQueryParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsQueryParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsQueryParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FilingsQueryParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsQueryParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsQueryParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FilingsQueryParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsQueryParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FilingsQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsQueryParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsQueryParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsQueryParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FilingsQueryParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsQueryParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsQueryParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FilingsQueryParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsQueryParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FilingsQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsQueryParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsQueryParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FilingsQueryParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsQueryParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsQueryParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsQueryParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsQueryParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FilingsQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsQueryParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsQueryParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsQueryParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsQueryParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsQueryParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsQueryParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsQueryParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsQueryParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsQueryParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilingsQueryParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilingsQueryParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilingsQueryParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
