% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnote_period_parameters.R
\docType{class}
\name{FootnotePeriodParameters}
\alias{FootnotePeriodParameters}
\title{FootnotePeriodParameters}
\format{
An \code{R6Class} generator object
}
\description{
FootnotePeriodParameters Class

FootnotePeriodParameters Class
}
\details{
Create a new FootnotePeriodParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{period}}{character \link{optional}}

\item{\code{year}}{integer \link{optional}}

\item{\code{endYear}}{integer \link{optional}}

\item{\code{endPeriod}}{character \link{optional}}

\item{\code{periodType}}{character \link{optional}}

\item{\code{useFiscalPeriod}}{character \link{optional}}

\item{\code{dateRange}}{\link{DateRange} \link{optional}}

\item{\code{allHistory}}{character \link{optional}}

\item{\code{updateDate}}{character \link{optional}}

\item{\code{updatedFrom}}{character \link{optional}}

\item{\code{asOriginallyReported}}{character \link{optional}}

\item{\code{accessionID}}{integer \link{optional}}

\item{\code{filingID}}{integer \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FootnotePeriodParameters-new}{\code{FootnotePeriodParameters$new()}}
\item \href{#method-FootnotePeriodParameters-toJSON}{\code{FootnotePeriodParameters$toJSON()}}
\item \href{#method-FootnotePeriodParameters-fromJSON}{\code{FootnotePeriodParameters$fromJSON()}}
\item \href{#method-FootnotePeriodParameters-toJSONString}{\code{FootnotePeriodParameters$toJSONString()}}
\item \href{#method-FootnotePeriodParameters-fromJSONString}{\code{FootnotePeriodParameters$fromJSONString()}}
\item \href{#method-FootnotePeriodParameters-validateJSON}{\code{FootnotePeriodParameters$validateJSON()}}
\item \href{#method-FootnotePeriodParameters-toString}{\code{FootnotePeriodParameters$toString()}}
\item \href{#method-FootnotePeriodParameters-isValid}{\code{FootnotePeriodParameters$isValid()}}
\item \href{#method-FootnotePeriodParameters-getInvalidFields}{\code{FootnotePeriodParameters$getInvalidFields()}}
\item \href{#method-FootnotePeriodParameters-print}{\code{FootnotePeriodParameters$print()}}
\item \href{#method-FootnotePeriodParameters-clone}{\code{FootnotePeriodParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnotePeriodParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-FootnotePeriodParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FootnotePeriodParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnotePeriodParameters$new(
  period = NULL,
  year = NULL,
  endYear = NULL,
  endPeriod = NULL,
  periodType = NULL,
  useFiscalPeriod = NULL,
  dateRange = NULL,
  allHistory = NULL,
  updateDate = NULL,
  updatedFrom = NULL,
  asOriginallyReported = NULL,
  accessionID = NULL,
  filingID = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{period}}{period}

\item{\code{year}}{year}

\item{\code{endYear}}{endYear}

\item{\code{endPeriod}}{endPeriod}

\item{\code{periodType}}{periodType}

\item{\code{useFiscalPeriod}}{useFiscalPeriod}

\item{\code{dateRange}}{dateRange}

\item{\code{allHistory}}{allHistory}

\item{\code{updateDate}}{updateDate}

\item{\code{updatedFrom}}{updatedFrom}

\item{\code{asOriginallyReported}}{asOriginallyReported}

\item{\code{accessionID}}{accessionID}

\item{\code{filingID}}{filingID}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnotePeriodParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FootnotePeriodParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnotePeriodParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FootnotePeriodParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnotePeriodParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FootnotePeriodParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FootnotePeriodParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnotePeriodParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FootnotePeriodParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnotePeriodParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FootnotePeriodParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnotePeriodParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FootnotePeriodParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnotePeriodParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FootnotePeriodParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FootnotePeriodParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnotePeriodParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FootnotePeriodParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnotePeriodParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FootnotePeriodParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FootnotePeriodParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnotePeriodParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnotePeriodParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FootnotePeriodParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnotePeriodParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FootnotePeriodParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnotePeriodParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FootnotePeriodParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnotePeriodParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnotePeriodParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FootnotePeriodParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnotePeriodParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnotePeriodParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-FootnotePeriodParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnotePeriodParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnotePeriodParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FootnotePeriodParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnotePeriodParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
