% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnote_query_parameters.R
\docType{class}
\name{FootnoteQueryParameters}
\alias{FootnoteQueryParameters}
\title{FootnoteQueryParameters}
\format{
An \code{R6Class} generator object
}
\description{
FootnoteQueryParameters Class

FootnoteQueryParameters Class
}
\details{
Create a new FootnoteQueryParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fullTextQuery}}{character \link{optional}}

\item{\code{arcRoles}}{list(integer) \link{optional}}

\item{\code{footnoteTag}}{character \link{optional}}

\item{\code{startEntityID}}{integer \link{optional}}

\item{\code{startOffset}}{integer \link{optional}}

\item{\code{accessionID}}{integer \link{optional}}

\item{\code{allFootnotes}}{character \link{optional}}

\item{\code{includeRelatedItems}}{character \link{optional}}

\item{\code{fullTextSearchAllTimePeriods}}{character \link{optional}}

\item{\code{fullTextSearchRestrictToMetricLabels}}{character \link{optional}}

\item{\code{normalizedFaceMetric}}{character \link{optional}}

\item{\code{normalizedFootnoteMetric}}{character \link{optional}}

\item{\code{normalizedBreakoutMetric}}{character \link{optional}}

\item{\code{ByDocumentID}}{character \link{optional}}

\item{\code{XBRLNumericTag}}{character \link{optional}}

\item{\code{FactID}}{integer \link{optional}}

\item{\code{blobID}}{character \link{optional}}

\item{\code{nonXBRLFactIdentifier}}{integer \link{optional}}

\item{\code{nonXBRLExtractTag}}{character \link{optional}}

\item{\code{XBRLFactIdentifier}}{integer \link{optional}}

\item{\code{userSubmittedURL}}{character \link{optional}}

\item{\code{subDivide}}{character \link{optional}}

\item{\code{publicDisclosureID}}{character \link{optional}}

\item{\code{historyFactID}}{integer \link{optional}}

\item{\code{disclosureName}}{character \link{optional}}

\item{\code{disclosureNames}}{list(character) \link{optional}}

\item{\code{limit}}{Maximum number of results to get integer \link{optional}}

\item{\code{footnoteType}}{character \link{optional}}

\item{\code{NetworkID}}{Search by XBRL Network ID integer \link{optional}}

\item{\code{AllTextBlocks}}{Get all of the XBRL text blocks character \link{optional}}

\item{\code{AllMDASections}}{Get all MDA sections for a single filing character \link{optional}}

\item{\code{FilingID}}{For non-XBRL proxy filings, etc.  Specifying this will get you all the documents for the company/period that the document is associated with in the SECFiling table. integer \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FootnoteQueryParameters-new}{\code{FootnoteQueryParameters$new()}}
\item \href{#method-FootnoteQueryParameters-toJSON}{\code{FootnoteQueryParameters$toJSON()}}
\item \href{#method-FootnoteQueryParameters-fromJSON}{\code{FootnoteQueryParameters$fromJSON()}}
\item \href{#method-FootnoteQueryParameters-toJSONString}{\code{FootnoteQueryParameters$toJSONString()}}
\item \href{#method-FootnoteQueryParameters-fromJSONString}{\code{FootnoteQueryParameters$fromJSONString()}}
\item \href{#method-FootnoteQueryParameters-validateJSON}{\code{FootnoteQueryParameters$validateJSON()}}
\item \href{#method-FootnoteQueryParameters-toString}{\code{FootnoteQueryParameters$toString()}}
\item \href{#method-FootnoteQueryParameters-isValid}{\code{FootnoteQueryParameters$isValid()}}
\item \href{#method-FootnoteQueryParameters-getInvalidFields}{\code{FootnoteQueryParameters$getInvalidFields()}}
\item \href{#method-FootnoteQueryParameters-print}{\code{FootnoteQueryParameters$print()}}
\item \href{#method-FootnoteQueryParameters-clone}{\code{FootnoteQueryParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FootnoteQueryParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryParameters$new(
  fullTextQuery = NULL,
  arcRoles = NULL,
  footnoteTag = NULL,
  startEntityID = NULL,
  startOffset = NULL,
  accessionID = NULL,
  allFootnotes = NULL,
  includeRelatedItems = NULL,
  fullTextSearchAllTimePeriods = NULL,
  fullTextSearchRestrictToMetricLabels = NULL,
  normalizedFaceMetric = NULL,
  normalizedFootnoteMetric = NULL,
  normalizedBreakoutMetric = NULL,
  ByDocumentID = NULL,
  XBRLNumericTag = NULL,
  FactID = NULL,
  blobID = NULL,
  nonXBRLFactIdentifier = NULL,
  nonXBRLExtractTag = NULL,
  XBRLFactIdentifier = NULL,
  userSubmittedURL = NULL,
  subDivide = NULL,
  publicDisclosureID = NULL,
  historyFactID = NULL,
  disclosureName = NULL,
  disclosureNames = NULL,
  limit = NULL,
  footnoteType = NULL,
  NetworkID = NULL,
  AllTextBlocks = NULL,
  AllMDASections = NULL,
  FilingID = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fullTextQuery}}{fullTextQuery}

\item{\code{arcRoles}}{arcRoles}

\item{\code{footnoteTag}}{footnoteTag}

\item{\code{startEntityID}}{startEntityID}

\item{\code{startOffset}}{startOffset}

\item{\code{accessionID}}{accessionID}

\item{\code{allFootnotes}}{allFootnotes}

\item{\code{includeRelatedItems}}{includeRelatedItems}

\item{\code{fullTextSearchAllTimePeriods}}{fullTextSearchAllTimePeriods}

\item{\code{fullTextSearchRestrictToMetricLabels}}{fullTextSearchRestrictToMetricLabels}

\item{\code{normalizedFaceMetric}}{normalizedFaceMetric}

\item{\code{normalizedFootnoteMetric}}{normalizedFootnoteMetric}

\item{\code{normalizedBreakoutMetric}}{normalizedBreakoutMetric}

\item{\code{ByDocumentID}}{ByDocumentID}

\item{\code{XBRLNumericTag}}{XBRLNumericTag}

\item{\code{FactID}}{FactID}

\item{\code{blobID}}{blobID}

\item{\code{nonXBRLFactIdentifier}}{nonXBRLFactIdentifier}

\item{\code{nonXBRLExtractTag}}{nonXBRLExtractTag}

\item{\code{XBRLFactIdentifier}}{XBRLFactIdentifier}

\item{\code{userSubmittedURL}}{userSubmittedURL}

\item{\code{subDivide}}{subDivide}

\item{\code{publicDisclosureID}}{publicDisclosureID}

\item{\code{historyFactID}}{historyFactID}

\item{\code{disclosureName}}{disclosureName}

\item{\code{disclosureNames}}{disclosureNames}

\item{\code{limit}}{Maximum number of results to get}

\item{\code{footnoteType}}{footnoteType}

\item{\code{NetworkID}}{Search by XBRL Network ID}

\item{\code{AllTextBlocks}}{Get all of the XBRL text blocks}

\item{\code{AllMDASections}}{Get all MDA sections for a single filing}

\item{\code{FilingID}}{For non-XBRL proxy filings, etc.  Specifying this will get you all the documents for the company/period that the document is associated with in the SECFiling table.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FootnoteQueryParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FootnoteQueryParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FootnoteQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FootnoteQueryParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FootnoteQueryParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FootnoteQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FootnoteQueryParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FootnoteQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
