% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnote_query_response.R
\docType{class}
\name{FootnoteQueryResponse}
\alias{FootnoteQueryResponse}
\title{FootnoteQueryResponse}
\format{
An \code{R6Class} generator object
}
\description{
FootnoteQueryResponse Class

FootnoteQueryResponse Class
}
\details{
Create a new FootnoteQueryResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{footnotes}}{Matching disclosures list(\link{TextFootnote}) \link{optional}}

\item{\code{nextGroupStartEntityID}}{integer \link{optional}}

\item{\code{nextGroupStartOffset}}{integer \link{optional}}

\item{\code{moreResults}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FootnoteQueryResponse-new}{\code{FootnoteQueryResponse$new()}}
\item \href{#method-FootnoteQueryResponse-toJSON}{\code{FootnoteQueryResponse$toJSON()}}
\item \href{#method-FootnoteQueryResponse-fromJSON}{\code{FootnoteQueryResponse$fromJSON()}}
\item \href{#method-FootnoteQueryResponse-toJSONString}{\code{FootnoteQueryResponse$toJSONString()}}
\item \href{#method-FootnoteQueryResponse-fromJSONString}{\code{FootnoteQueryResponse$fromJSONString()}}
\item \href{#method-FootnoteQueryResponse-validateJSON}{\code{FootnoteQueryResponse$validateJSON()}}
\item \href{#method-FootnoteQueryResponse-toString}{\code{FootnoteQueryResponse$toString()}}
\item \href{#method-FootnoteQueryResponse-isValid}{\code{FootnoteQueryResponse$isValid()}}
\item \href{#method-FootnoteQueryResponse-getInvalidFields}{\code{FootnoteQueryResponse$getInvalidFields()}}
\item \href{#method-FootnoteQueryResponse-print}{\code{FootnoteQueryResponse$print()}}
\item \href{#method-FootnoteQueryResponse-clone}{\code{FootnoteQueryResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FootnoteQueryResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryResponse$new(
  footnotes = NULL,
  nextGroupStartEntityID = NULL,
  nextGroupStartOffset = NULL,
  moreResults = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{footnotes}}{Matching disclosures}

\item{\code{nextGroupStartEntityID}}{nextGroupStartEntityID}

\item{\code{nextGroupStartOffset}}{nextGroupStartOffset}

\item{\code{moreResults}}{moreResults}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FootnoteQueryResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FootnoteQueryResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FootnoteQueryResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FootnoteQueryResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FootnoteQueryResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FootnoteQueryResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FootnoteQueryResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FootnoteQueryResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteQueryResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteQueryResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteQueryResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
