% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnote_search_parameters.R
\docType{class}
\name{FootnoteSearchParameters}
\alias{FootnoteSearchParameters}
\title{FootnoteSearchParameters}
\format{
An \code{R6Class} generator object
}
\description{
FootnoteSearchParameters Class

FootnoteSearchParameters Class
}
\details{
Create a new FootnoteSearchParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pageParameters}}{\link{FootnoteQueryParameters} \link{optional}}

\item{\code{periodParameters}}{\link{FootnotePeriodParameters} \link{optional}}

\item{\code{companiesParameters}}{\link{CompaniesParameters} \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FootnoteSearchParameters-new}{\code{FootnoteSearchParameters$new()}}
\item \href{#method-FootnoteSearchParameters-toJSON}{\code{FootnoteSearchParameters$toJSON()}}
\item \href{#method-FootnoteSearchParameters-fromJSON}{\code{FootnoteSearchParameters$fromJSON()}}
\item \href{#method-FootnoteSearchParameters-toJSONString}{\code{FootnoteSearchParameters$toJSONString()}}
\item \href{#method-FootnoteSearchParameters-fromJSONString}{\code{FootnoteSearchParameters$fromJSONString()}}
\item \href{#method-FootnoteSearchParameters-validateJSON}{\code{FootnoteSearchParameters$validateJSON()}}
\item \href{#method-FootnoteSearchParameters-toString}{\code{FootnoteSearchParameters$toString()}}
\item \href{#method-FootnoteSearchParameters-isValid}{\code{FootnoteSearchParameters$isValid()}}
\item \href{#method-FootnoteSearchParameters-getInvalidFields}{\code{FootnoteSearchParameters$getInvalidFields()}}
\item \href{#method-FootnoteSearchParameters-print}{\code{FootnoteSearchParameters$print()}}
\item \href{#method-FootnoteSearchParameters-clone}{\code{FootnoteSearchParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteSearchParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteSearchParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FootnoteSearchParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteSearchParameters$new(
  pageParameters = NULL,
  periodParameters = NULL,
  companiesParameters = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pageParameters}}{pageParameters}

\item{\code{periodParameters}}{periodParameters}

\item{\code{companiesParameters}}{companiesParameters}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteSearchParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteSearchParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteSearchParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FootnoteSearchParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteSearchParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteSearchParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FootnoteSearchParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteSearchParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FootnoteSearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteSearchParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteSearchParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteSearchParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FootnoteSearchParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteSearchParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteSearchParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FootnoteSearchParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteSearchParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FootnoteSearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteSearchParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteSearchParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FootnoteSearchParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteSearchParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteSearchParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteSearchParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteSearchParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FootnoteSearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteSearchParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteSearchParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteSearchParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteSearchParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteSearchParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteSearchParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteSearchParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteSearchParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteSearchParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FootnoteSearchParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FootnoteSearchParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FootnoteSearchParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
