% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form13_f_query_piece.R
\docType{class}
\name{Form13FQueryPiece}
\alias{Form13FQueryPiece}
\title{Form13FQueryPiece}
\format{
An \code{R6Class} generator object
}
\description{
Form13FQueryPiece Class

Form13FQueryPiece Class
}
\details{
Create a new Form13FQueryPiece
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parameter}}{character \link{optional}}

\item{\code{Operator}}{integer \link{optional}}

\item{\code{value}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Form13FQueryPiece-new}{\code{Form13FQueryPiece$new()}}
\item \href{#method-Form13FQueryPiece-toJSON}{\code{Form13FQueryPiece$toJSON()}}
\item \href{#method-Form13FQueryPiece-fromJSON}{\code{Form13FQueryPiece$fromJSON()}}
\item \href{#method-Form13FQueryPiece-toJSONString}{\code{Form13FQueryPiece$toJSONString()}}
\item \href{#method-Form13FQueryPiece-fromJSONString}{\code{Form13FQueryPiece$fromJSONString()}}
\item \href{#method-Form13FQueryPiece-validateJSON}{\code{Form13FQueryPiece$validateJSON()}}
\item \href{#method-Form13FQueryPiece-toString}{\code{Form13FQueryPiece$toString()}}
\item \href{#method-Form13FQueryPiece-isValid}{\code{Form13FQueryPiece$isValid()}}
\item \href{#method-Form13FQueryPiece-getInvalidFields}{\code{Form13FQueryPiece$getInvalidFields()}}
\item \href{#method-Form13FQueryPiece-print}{\code{Form13FQueryPiece$print()}}
\item \href{#method-Form13FQueryPiece-clone}{\code{Form13FQueryPiece$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Form13FQueryPiece-new"></a>}}
\if{latex}{\out{\hypertarget{method-Form13FQueryPiece-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Form13FQueryPiece class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Form13FQueryPiece$new(parameter = NULL, Operator = NULL, value = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameter}}{parameter}

\item{\code{Operator}}{Operator}

\item{\code{value}}{value}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Form13FQueryPiece-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Form13FQueryPiece-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Form13FQueryPiece$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Form13FQueryPiece in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Form13FQueryPiece-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Form13FQueryPiece-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Form13FQueryPiece
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Form13FQueryPiece$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Form13FQueryPiece
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Form13FQueryPiece-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Form13FQueryPiece-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Form13FQueryPiece$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Form13FQueryPiece in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Form13FQueryPiece-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Form13FQueryPiece-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Form13FQueryPiece
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Form13FQueryPiece$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Form13FQueryPiece
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Form13FQueryPiece-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Form13FQueryPiece-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Form13FQueryPiece and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Form13FQueryPiece$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Form13FQueryPiece-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Form13FQueryPiece-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Form13FQueryPiece$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Form13FQueryPiece
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Form13FQueryPiece-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Form13FQueryPiece-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Form13FQueryPiece$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Form13FQueryPiece-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Form13FQueryPiece-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Form13FQueryPiece$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Form13FQueryPiece-print"></a>}}
\if{latex}{\out{\hypertarget{method-Form13FQueryPiece-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Form13FQueryPiece$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Form13FQueryPiece-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Form13FQueryPiece-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Form13FQueryPiece$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
