% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_companies.R
\name{get_available_companies}
\alias{get_available_companies}
\title{Get Available Compa ies}
\usage{
get_available_companies(index)
}
\value{
data.table
}
\description{
Function to get Calcbench available companies
}
