% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_metrics.R
\name{get_available_metrics}
\alias{get_available_metrics}
\title{Get Available Metrics}
\usage{
get_available_metrics()
}
\value{
data.table
}
\description{
Function to get Calcbench available metrics
}
