% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filings.R
\name{get_filings}
\alias{get_filings}
\title{Get Standardized Data}
\usage{
get_filings(
  companies,
  entire_universe = FALSE,
  all_history = TRUE,
  year = NULL,
  end_year = NULL,
  period = NULL,
  end_period = NULL,
  period_type = NULL,
  use_fiscal_period = NULL,
  date_range = NULL,
  update_date = NULL,
  updated_from = NULL,
  as_originally_reported = FALSE,
  accession_id = NULL,
  filing_id = NULL,
  include_performance = TRUE,
  include_non_xbrl = TRUE,
  include_press_release_and_proxies = TRUE,
  include_xbrl = TRUE
)
}
\arguments{
\item{companies}{list or chr (list("msft") or "msft")}

\item{entire_universe}{boolean}

\item{all_history}{boolean}

\item{year}{integer}

\item{end_year}{integer}

\item{period}{integer}

\item{end_period}{integer}

\item{accession_id}{integer}

\item{filing_id}{integer}

\item{include_xbrl}{}
}
\value{
data.table
}
\description{
Function taking companies, metrics and history and returning standardized data
}
