% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standardized_data.R
\name{get_standardized_data}
\alias{get_standardized_data}
\title{Get Standardized Data}
\usage{
get_standardized_data(metrics, companies, all_history)
}
\arguments{
\item{metrics}{list of chr ie: list("revenue")}

\item{companies}{list of chr ie: list("msft")}

\item{all_history}{boolean}

\item{api_client}{Calcbench api_client object}
}
\value{
data.table
}
\description{
Function taking companies, metrics and history and returning standardized data
}
