% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intangible_assets_category.R
\docType{class}
\name{IntangibleAssetsCategory}
\alias{IntangibleAssetsCategory}
\title{IntangibleAssetsCategory}
\format{
An \code{R6Class} generator object
}
\description{
IntangibleAssetsCategory Class

IntangibleAssetsCategory Class
}
\details{
Create a new IntangibleAssetsCategory
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{category}}{character \link{optional}}

\item{\code{value}}{numeric \link{optional}}

\item{\code{useful_life_upper_range}}{numeric \link{optional}}

\item{\code{useful_life_lower_range}}{numeric \link{optional}}

\item{\code{trace_facts}}{list(\link{TraceData}) \link{optional}}

\item{\code{format_type}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IntangibleAssetsCategory-new}{\code{IntangibleAssetsCategory$new()}}
\item \href{#method-IntangibleAssetsCategory-toJSON}{\code{IntangibleAssetsCategory$toJSON()}}
\item \href{#method-IntangibleAssetsCategory-fromJSON}{\code{IntangibleAssetsCategory$fromJSON()}}
\item \href{#method-IntangibleAssetsCategory-toJSONString}{\code{IntangibleAssetsCategory$toJSONString()}}
\item \href{#method-IntangibleAssetsCategory-fromJSONString}{\code{IntangibleAssetsCategory$fromJSONString()}}
\item \href{#method-IntangibleAssetsCategory-validateJSON}{\code{IntangibleAssetsCategory$validateJSON()}}
\item \href{#method-IntangibleAssetsCategory-toString}{\code{IntangibleAssetsCategory$toString()}}
\item \href{#method-IntangibleAssetsCategory-isValid}{\code{IntangibleAssetsCategory$isValid()}}
\item \href{#method-IntangibleAssetsCategory-getInvalidFields}{\code{IntangibleAssetsCategory$getInvalidFields()}}
\item \href{#method-IntangibleAssetsCategory-print}{\code{IntangibleAssetsCategory$print()}}
\item \href{#method-IntangibleAssetsCategory-clone}{\code{IntangibleAssetsCategory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntangibleAssetsCategory-new"></a>}}
\if{latex}{\out{\hypertarget{method-IntangibleAssetsCategory-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new IntangibleAssetsCategory class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntangibleAssetsCategory$new(
  category = NULL,
  value = NULL,
  useful_life_upper_range = NULL,
  useful_life_lower_range = NULL,
  trace_facts = NULL,
  format_type = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{category}

\item{\code{value}}{value}

\item{\code{useful_life_upper_range}}{useful_life_upper_range}

\item{\code{useful_life_lower_range}}{useful_life_lower_range}

\item{\code{trace_facts}}{trace_facts}

\item{\code{format_type}}{format_type}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntangibleAssetsCategory-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-IntangibleAssetsCategory-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntangibleAssetsCategory$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
IntangibleAssetsCategory in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntangibleAssetsCategory-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-IntangibleAssetsCategory-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of IntangibleAssetsCategory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntangibleAssetsCategory$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of IntangibleAssetsCategory
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntangibleAssetsCategory-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-IntangibleAssetsCategory-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntangibleAssetsCategory$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
IntangibleAssetsCategory in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntangibleAssetsCategory-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-IntangibleAssetsCategory-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of IntangibleAssetsCategory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntangibleAssetsCategory$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of IntangibleAssetsCategory
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntangibleAssetsCategory-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-IntangibleAssetsCategory-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to IntangibleAssetsCategory and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntangibleAssetsCategory$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntangibleAssetsCategory-toString"></a>}}
\if{latex}{\out{\hypertarget{method-IntangibleAssetsCategory-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntangibleAssetsCategory$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of IntangibleAssetsCategory
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntangibleAssetsCategory-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-IntangibleAssetsCategory-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntangibleAssetsCategory$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntangibleAssetsCategory-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-IntangibleAssetsCategory-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntangibleAssetsCategory$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntangibleAssetsCategory-print"></a>}}
\if{latex}{\out{\hypertarget{method-IntangibleAssetsCategory-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntangibleAssetsCategory$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntangibleAssetsCategory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IntangibleAssetsCategory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntangibleAssetsCategory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
