% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapped_data_point.R
\docType{class}
\name{MappedDataPoint}
\alias{MappedDataPoint}
\title{MappedDataPoint}
\format{
An \code{R6Class} generator object
}
\description{
MappedDataPoint Class

MappedDataPoint Class
}
\details{
Create a new MappedDataPoint
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{standardized_id}}{integer \link{optional}}

\item{\code{metric}}{character \link{optional}}

\item{\code{value}}{object \link{optional}}

\item{\code{calendar_year}}{integer \link{optional}}

\item{\code{calendar_period}}{integer \link{optional}}

\item{\code{fiscal_year}}{integer \link{optional}}

\item{\code{fiscal_period}}{integer \link{optional}}

\item{\code{trace_facts}}{list(\link{TraceData}) \link{optional}}

\item{\code{ticker}}{character \link{optional}}

\item{\code{CIK}}{character \link{optional}}

\item{\code{calcbench_entity_id}}{integer \link{optional}}

\item{\code{filing_type}}{character \link{optional}}

\item{\code{preliminary}}{character \link{optional}}

\item{\code{filing_accession_number}}{character \link{optional}}

\item{\code{trace_url}}{character \link{optional}}

\item{\code{revision_number}}{integer \link{optional}}

\item{\code{XBRL}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MappedDataPoint-new}{\code{MappedDataPoint$new()}}
\item \href{#method-MappedDataPoint-toJSON}{\code{MappedDataPoint$toJSON()}}
\item \href{#method-MappedDataPoint-fromJSON}{\code{MappedDataPoint$fromJSON()}}
\item \href{#method-MappedDataPoint-toJSONString}{\code{MappedDataPoint$toJSONString()}}
\item \href{#method-MappedDataPoint-fromJSONString}{\code{MappedDataPoint$fromJSONString()}}
\item \href{#method-MappedDataPoint-validateJSON}{\code{MappedDataPoint$validateJSON()}}
\item \href{#method-MappedDataPoint-toString}{\code{MappedDataPoint$toString()}}
\item \href{#method-MappedDataPoint-isValid}{\code{MappedDataPoint$isValid()}}
\item \href{#method-MappedDataPoint-getInvalidFields}{\code{MappedDataPoint$getInvalidFields()}}
\item \href{#method-MappedDataPoint-print}{\code{MappedDataPoint$print()}}
\item \href{#method-MappedDataPoint-clone}{\code{MappedDataPoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataPoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataPoint-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new MappedDataPoint class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataPoint$new(
  standardized_id = NULL,
  metric = NULL,
  value = NULL,
  calendar_year = NULL,
  calendar_period = NULL,
  fiscal_year = NULL,
  fiscal_period = NULL,
  trace_facts = NULL,
  ticker = NULL,
  CIK = NULL,
  calcbench_entity_id = NULL,
  filing_type = NULL,
  preliminary = NULL,
  filing_accession_number = NULL,
  trace_url = NULL,
  revision_number = NULL,
  XBRL = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{standardized_id}}{standardized_id}

\item{\code{metric}}{metric}

\item{\code{value}}{value}

\item{\code{calendar_year}}{calendar_year}

\item{\code{calendar_period}}{calendar_period}

\item{\code{fiscal_year}}{fiscal_year}

\item{\code{fiscal_period}}{fiscal_period}

\item{\code{trace_facts}}{trace_facts}

\item{\code{ticker}}{ticker}

\item{\code{CIK}}{CIK}

\item{\code{calcbench_entity_id}}{calcbench_entity_id}

\item{\code{filing_type}}{filing_type}

\item{\code{preliminary}}{preliminary}

\item{\code{filing_accession_number}}{filing_accession_number}

\item{\code{trace_url}}{trace_url}

\item{\code{revision_number}}{revision_number}

\item{\code{XBRL}}{XBRL}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataPoint-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataPoint-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataPoint$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MappedDataPoint in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataPoint-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataPoint-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of MappedDataPoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataPoint$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MappedDataPoint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataPoint-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataPoint-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataPoint$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MappedDataPoint in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataPoint-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataPoint-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of MappedDataPoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataPoint$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MappedDataPoint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataPoint-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataPoint-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to MappedDataPoint and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataPoint$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataPoint-toString"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataPoint-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataPoint$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of MappedDataPoint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataPoint-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataPoint-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataPoint$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataPoint-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataPoint-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataPoint$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataPoint-print"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataPoint-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataPoint$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataPoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataPoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataPoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
