% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapped_data_search_parameters.R
\docType{class}
\name{MappedDataSearchParameters}
\alias{MappedDataSearchParameters}
\title{MappedDataSearchParameters}
\format{
An \code{R6Class} generator object
}
\description{
MappedDataSearchParameters Class

MappedDataSearchParameters Class
}
\details{
Create a new MappedDataSearchParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pageParameters}}{\link{StandardizedQueryParameters} \link{optional}}

\item{\code{periodParameters}}{\link{PeriodParameters} \link{optional}}

\item{\code{companiesParameters}}{\link{CompaniesParameters} \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MappedDataSearchParameters-new}{\code{MappedDataSearchParameters$new()}}
\item \href{#method-MappedDataSearchParameters-toJSON}{\code{MappedDataSearchParameters$toJSON()}}
\item \href{#method-MappedDataSearchParameters-fromJSON}{\code{MappedDataSearchParameters$fromJSON()}}
\item \href{#method-MappedDataSearchParameters-toJSONString}{\code{MappedDataSearchParameters$toJSONString()}}
\item \href{#method-MappedDataSearchParameters-fromJSONString}{\code{MappedDataSearchParameters$fromJSONString()}}
\item \href{#method-MappedDataSearchParameters-validateJSON}{\code{MappedDataSearchParameters$validateJSON()}}
\item \href{#method-MappedDataSearchParameters-toString}{\code{MappedDataSearchParameters$toString()}}
\item \href{#method-MappedDataSearchParameters-isValid}{\code{MappedDataSearchParameters$isValid()}}
\item \href{#method-MappedDataSearchParameters-getInvalidFields}{\code{MappedDataSearchParameters$getInvalidFields()}}
\item \href{#method-MappedDataSearchParameters-print}{\code{MappedDataSearchParameters$print()}}
\item \href{#method-MappedDataSearchParameters-clone}{\code{MappedDataSearchParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataSearchParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataSearchParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new MappedDataSearchParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataSearchParameters$new(
  pageParameters = NULL,
  periodParameters = NULL,
  companiesParameters = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pageParameters}}{pageParameters}

\item{\code{periodParameters}}{periodParameters}

\item{\code{companiesParameters}}{companiesParameters}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataSearchParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataSearchParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataSearchParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MappedDataSearchParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataSearchParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataSearchParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of MappedDataSearchParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataSearchParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MappedDataSearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataSearchParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataSearchParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataSearchParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MappedDataSearchParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataSearchParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataSearchParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of MappedDataSearchParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataSearchParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MappedDataSearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataSearchParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataSearchParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to MappedDataSearchParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataSearchParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataSearchParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataSearchParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataSearchParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of MappedDataSearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataSearchParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataSearchParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataSearchParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataSearchParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataSearchParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataSearchParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataSearchParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataSearchParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataSearchParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataSearchParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataSearchParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataSearchParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
