% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_groups.R
\docType{class}
\name{MetricGroups}
\alias{MetricGroups}
\title{MetricGroups}
\format{
An \code{R6Class} generator object
}
\description{
MetricGroups Class

MetricGroups Class
}
\details{
Create a new MetricGroups
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{face}}{list(\link{StandardizedMetricDefinition}) \link{optional}}

\item{\code{footnote}}{list(\link{StandardizedFootnoteMetricDefinition}) \link{optional}}

\item{\code{ratio}}{list(\link{StandardizedMetricDefinition}) \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MetricGroups-new}{\code{MetricGroups$new()}}
\item \href{#method-MetricGroups-toJSON}{\code{MetricGroups$toJSON()}}
\item \href{#method-MetricGroups-fromJSON}{\code{MetricGroups$fromJSON()}}
\item \href{#method-MetricGroups-toJSONString}{\code{MetricGroups$toJSONString()}}
\item \href{#method-MetricGroups-fromJSONString}{\code{MetricGroups$fromJSONString()}}
\item \href{#method-MetricGroups-validateJSON}{\code{MetricGroups$validateJSON()}}
\item \href{#method-MetricGroups-toString}{\code{MetricGroups$toString()}}
\item \href{#method-MetricGroups-isValid}{\code{MetricGroups$isValid()}}
\item \href{#method-MetricGroups-getInvalidFields}{\code{MetricGroups$getInvalidFields()}}
\item \href{#method-MetricGroups-print}{\code{MetricGroups$print()}}
\item \href{#method-MetricGroups-clone}{\code{MetricGroups$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricGroups-new"></a>}}
\if{latex}{\out{\hypertarget{method-MetricGroups-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new MetricGroups class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricGroups$new(face = NULL, footnote = NULL, ratio = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{face}}{face}

\item{\code{footnote}}{footnote}

\item{\code{ratio}}{ratio}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricGroups-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MetricGroups-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricGroups$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MetricGroups in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricGroups-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MetricGroups-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of MetricGroups
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricGroups$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MetricGroups
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricGroups-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MetricGroups-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricGroups$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MetricGroups in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricGroups-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MetricGroups-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of MetricGroups
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricGroups$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MetricGroups
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricGroups-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MetricGroups-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to MetricGroups and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricGroups$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricGroups-toString"></a>}}
\if{latex}{\out{\hypertarget{method-MetricGroups-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricGroups$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of MetricGroups
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricGroups-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-MetricGroups-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricGroups$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricGroups-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-MetricGroups-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricGroups$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricGroups-print"></a>}}
\if{latex}{\out{\hypertarget{method-MetricGroups-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricGroups$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricGroups-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MetricGroups-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricGroups$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
