% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_search_criteria.R
\docType{class}
\name{MetricSearchCriteria}
\alias{MetricSearchCriteria}
\title{MetricSearchCriteria}
\format{
An \code{R6Class} generator object
}
\description{
MetricSearchCriteria Class

MetricSearchCriteria Class
}
\details{
Create a new MetricSearchCriteria
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metrics}}{list(character) \link{optional}}

\item{\code{company_identifiers}}{list(character) \link{optional}}

\item{\code{start_year}}{integer \link{optional}}

\item{\code{start_period}}{integer \link{optional}}

\item{\code{end_year}}{integer \link{optional}}

\item{\code{end_period}}{integer \link{optional}}

\item{\code{fiscal_period}}{character \link{optional}}

\item{\code{calendar_year}}{integer \link{optional}}

\item{\code{calendar_period}}{character \link{optional}}

\item{\code{periods_back}}{integer \link{optional}}

\item{\code{include_trace}}{character \link{optional}}

\item{\code{entire_universe}}{character \link{optional}}

\item{\code{AccessionID}}{integer \link{optional}}

\item{\code{use_fiscal_periods}}{character \link{optional}}

\item{\code{as_originally_reported}}{character \link{optional}}

\item{\code{all_history}}{character \link{optional}}

\item{\code{PointInTime}}{character \link{optional}}

\item{\code{updateDate}}{character \link{optional}}

\item{\code{period_type}}{character \link{optional}}

\item{\code{Annual}}{character \link{optional}}

\item{\code{Quarterly}}{character \link{optional}}

\item{\code{AllowedPeriods}}{list(character) \link{optional}}

\item{\code{metricsDictionary}}{named list(character) \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MetricSearchCriteria-new}{\code{MetricSearchCriteria$new()}}
\item \href{#method-MetricSearchCriteria-toJSON}{\code{MetricSearchCriteria$toJSON()}}
\item \href{#method-MetricSearchCriteria-fromJSON}{\code{MetricSearchCriteria$fromJSON()}}
\item \href{#method-MetricSearchCriteria-toJSONString}{\code{MetricSearchCriteria$toJSONString()}}
\item \href{#method-MetricSearchCriteria-fromJSONString}{\code{MetricSearchCriteria$fromJSONString()}}
\item \href{#method-MetricSearchCriteria-validateJSON}{\code{MetricSearchCriteria$validateJSON()}}
\item \href{#method-MetricSearchCriteria-toString}{\code{MetricSearchCriteria$toString()}}
\item \href{#method-MetricSearchCriteria-isValid}{\code{MetricSearchCriteria$isValid()}}
\item \href{#method-MetricSearchCriteria-getInvalidFields}{\code{MetricSearchCriteria$getInvalidFields()}}
\item \href{#method-MetricSearchCriteria-print}{\code{MetricSearchCriteria$print()}}
\item \href{#method-MetricSearchCriteria-clone}{\code{MetricSearchCriteria$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricSearchCriteria-new"></a>}}
\if{latex}{\out{\hypertarget{method-MetricSearchCriteria-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new MetricSearchCriteria class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricSearchCriteria$new(
  metrics = NULL,
  company_identifiers = NULL,
  start_year = NULL,
  start_period = NULL,
  end_year = NULL,
  end_period = NULL,
  fiscal_period = NULL,
  calendar_year = NULL,
  calendar_period = NULL,
  periods_back = NULL,
  include_trace = NULL,
  entire_universe = NULL,
  AccessionID = NULL,
  use_fiscal_periods = NULL,
  as_originally_reported = NULL,
  all_history = NULL,
  PointInTime = NULL,
  updateDate = NULL,
  period_type = NULL,
  Annual = NULL,
  Quarterly = NULL,
  AllowedPeriods = NULL,
  metricsDictionary = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metrics}}{metrics}

\item{\code{company_identifiers}}{company_identifiers}

\item{\code{start_year}}{start_year}

\item{\code{start_period}}{start_period}

\item{\code{end_year}}{end_year}

\item{\code{end_period}}{end_period}

\item{\code{fiscal_period}}{fiscal_period}

\item{\code{calendar_year}}{calendar_year}

\item{\code{calendar_period}}{calendar_period}

\item{\code{periods_back}}{periods_back}

\item{\code{include_trace}}{include_trace}

\item{\code{entire_universe}}{entire_universe}

\item{\code{AccessionID}}{AccessionID}

\item{\code{use_fiscal_periods}}{use_fiscal_periods}

\item{\code{as_originally_reported}}{as_originally_reported}

\item{\code{all_history}}{all_history}

\item{\code{PointInTime}}{PointInTime}

\item{\code{updateDate}}{updateDate}

\item{\code{period_type}}{period_type}

\item{\code{Annual}}{Annual}

\item{\code{Quarterly}}{Quarterly}

\item{\code{AllowedPeriods}}{AllowedPeriods}

\item{\code{metricsDictionary}}{metricsDictionary}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricSearchCriteria-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MetricSearchCriteria-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricSearchCriteria$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MetricSearchCriteria in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricSearchCriteria-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MetricSearchCriteria-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of MetricSearchCriteria
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricSearchCriteria$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MetricSearchCriteria
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricSearchCriteria-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MetricSearchCriteria-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricSearchCriteria$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MetricSearchCriteria in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricSearchCriteria-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MetricSearchCriteria-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of MetricSearchCriteria
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricSearchCriteria$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MetricSearchCriteria
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricSearchCriteria-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MetricSearchCriteria-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to MetricSearchCriteria and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricSearchCriteria$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricSearchCriteria-toString"></a>}}
\if{latex}{\out{\hypertarget{method-MetricSearchCriteria-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricSearchCriteria$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of MetricSearchCriteria
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricSearchCriteria-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-MetricSearchCriteria-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricSearchCriteria$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricSearchCriteria-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-MetricSearchCriteria-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricSearchCriteria$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricSearchCriteria-print"></a>}}
\if{latex}{\out{\hypertarget{method-MetricSearchCriteria-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricSearchCriteria$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetricSearchCriteria-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MetricSearchCriteria-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricSearchCriteria$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
