% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nxbrl_query_piece.R
\docType{class}
\name{NONxbrlQueryPiece}
\alias{NONxbrlQueryPiece}
\title{NONxbrlQueryPiece}
\format{
An \code{R6Class} generator object
}
\description{
NONxbrlQueryPiece Class

NONxbrlQueryPiece Class
}
\details{
Create a new NONxbrlQueryPiece
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parameter}}{character \link{optional}}

\item{\code{Operator}}{character \link{optional}}

\item{\code{value}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NONxbrlQueryPiece-new}{\code{NONxbrlQueryPiece$new()}}
\item \href{#method-NONxbrlQueryPiece-toJSON}{\code{NONxbrlQueryPiece$toJSON()}}
\item \href{#method-NONxbrlQueryPiece-fromJSON}{\code{NONxbrlQueryPiece$fromJSON()}}
\item \href{#method-NONxbrlQueryPiece-toJSONString}{\code{NONxbrlQueryPiece$toJSONString()}}
\item \href{#method-NONxbrlQueryPiece-fromJSONString}{\code{NONxbrlQueryPiece$fromJSONString()}}
\item \href{#method-NONxbrlQueryPiece-validateJSON}{\code{NONxbrlQueryPiece$validateJSON()}}
\item \href{#method-NONxbrlQueryPiece-toString}{\code{NONxbrlQueryPiece$toString()}}
\item \href{#method-NONxbrlQueryPiece-isValid}{\code{NONxbrlQueryPiece$isValid()}}
\item \href{#method-NONxbrlQueryPiece-getInvalidFields}{\code{NONxbrlQueryPiece$getInvalidFields()}}
\item \href{#method-NONxbrlQueryPiece-print}{\code{NONxbrlQueryPiece$print()}}
\item \href{#method-NONxbrlQueryPiece-clone}{\code{NONxbrlQueryPiece$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NONxbrlQueryPiece-new"></a>}}
\if{latex}{\out{\hypertarget{method-NONxbrlQueryPiece-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new NONxbrlQueryPiece class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NONxbrlQueryPiece$new(parameter = NULL, Operator = NULL, value = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameter}}{parameter}

\item{\code{Operator}}{Operator}

\item{\code{value}}{value}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NONxbrlQueryPiece-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NONxbrlQueryPiece-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NONxbrlQueryPiece$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NONxbrlQueryPiece in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NONxbrlQueryPiece-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NONxbrlQueryPiece-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of NONxbrlQueryPiece
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NONxbrlQueryPiece$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of NONxbrlQueryPiece
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NONxbrlQueryPiece-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NONxbrlQueryPiece-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NONxbrlQueryPiece$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NONxbrlQueryPiece in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NONxbrlQueryPiece-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NONxbrlQueryPiece-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of NONxbrlQueryPiece
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NONxbrlQueryPiece$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of NONxbrlQueryPiece
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NONxbrlQueryPiece-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NONxbrlQueryPiece-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to NONxbrlQueryPiece and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NONxbrlQueryPiece$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NONxbrlQueryPiece-toString"></a>}}
\if{latex}{\out{\hypertarget{method-NONxbrlQueryPiece-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NONxbrlQueryPiece$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of NONxbrlQueryPiece
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NONxbrlQueryPiece-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-NONxbrlQueryPiece-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NONxbrlQueryPiece$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NONxbrlQueryPiece-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-NONxbrlQueryPiece-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NONxbrlQueryPiece$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NONxbrlQueryPiece-print"></a>}}
\if{latex}{\out{\hypertarget{method-NONxbrlQueryPiece-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NONxbrlQueryPiece$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NONxbrlQueryPiece-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NONxbrlQueryPiece-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NONxbrlQueryPiece$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
