% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_parameters.R
\docType{class}
\name{PeriodParameters}
\alias{PeriodParameters}
\title{PeriodParameters}
\format{
An \code{R6Class} generator object
}
\description{
PeriodParameters Class

PeriodParameters Class
}
\details{
Create a new PeriodParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{year}}{integer \link{optional}}

\item{\code{period}}{character \link{optional}}

\item{\code{endYear}}{integer \link{optional}}

\item{\code{endPeriod}}{character \link{optional}}

\item{\code{periodType}}{character \link{optional}}

\item{\code{useFiscalPeriod}}{character \link{optional}}

\item{\code{dateRange}}{\link{DateRange} \link{optional}}

\item{\code{allHistory}}{character \link{optional}}

\item{\code{updateDate}}{character \link{optional}}

\item{\code{updatedFrom}}{character \link{optional}}

\item{\code{asOriginallyReported}}{character \link{optional}}

\item{\code{accessionID}}{integer \link{optional}}

\item{\code{filingID}}{integer \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PeriodParameters-new}{\code{PeriodParameters$new()}}
\item \href{#method-PeriodParameters-toJSON}{\code{PeriodParameters$toJSON()}}
\item \href{#method-PeriodParameters-fromJSON}{\code{PeriodParameters$fromJSON()}}
\item \href{#method-PeriodParameters-toJSONString}{\code{PeriodParameters$toJSONString()}}
\item \href{#method-PeriodParameters-fromJSONString}{\code{PeriodParameters$fromJSONString()}}
\item \href{#method-PeriodParameters-validateJSON}{\code{PeriodParameters$validateJSON()}}
\item \href{#method-PeriodParameters-toString}{\code{PeriodParameters$toString()}}
\item \href{#method-PeriodParameters-isValid}{\code{PeriodParameters$isValid()}}
\item \href{#method-PeriodParameters-getInvalidFields}{\code{PeriodParameters$getInvalidFields()}}
\item \href{#method-PeriodParameters-print}{\code{PeriodParameters$print()}}
\item \href{#method-PeriodParameters-clone}{\code{PeriodParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PeriodParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodParameters$new(
  year = NULL,
  period = NULL,
  endYear = NULL,
  endPeriod = NULL,
  periodType = NULL,
  useFiscalPeriod = NULL,
  dateRange = NULL,
  allHistory = NULL,
  updateDate = NULL,
  updatedFrom = NULL,
  asOriginallyReported = NULL,
  accessionID = NULL,
  filingID = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{year}}{year}

\item{\code{period}}{period}

\item{\code{endYear}}{endYear}

\item{\code{endPeriod}}{endPeriod}

\item{\code{periodType}}{periodType}

\item{\code{useFiscalPeriod}}{useFiscalPeriod}

\item{\code{dateRange}}{dateRange}

\item{\code{allHistory}}{allHistory}

\item{\code{updateDate}}{updateDate}

\item{\code{updatedFrom}}{updatedFrom}

\item{\code{asOriginallyReported}}{asOriginallyReported}

\item{\code{accessionID}}{accessionID}

\item{\code{filingID}}{filingID}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PeriodParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PeriodParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PeriodParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PeriodParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PeriodParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PeriodParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PeriodParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PeriodParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
