% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchase_price_allocation.R
\docType{class}
\name{PurchasePriceAllocation}
\alias{PurchasePriceAllocation}
\title{PurchasePriceAllocation}
\format{
An \code{R6Class} generator object
}
\description{
Standardized M and A data

PurchasePriceAllocation Class
}
\details{
Create a new PurchasePriceAllocation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{acquisition_date}}{character \link{optional}}

\item{\code{date_reported}}{character \link{optional}}

\item{\code{date_originally_reported}}{character \link{optional}}

\item{\code{target}}{The company being acquired character \link{optional}}

\item{\code{enterprise_value}}{Purchase price + Debt - Cash numeric \link{optional}}

\item{\code{parent_company}}{The acquiring company character \link{optional}}

\item{\code{parent_company_ticker}}{character \link{optional}}

\item{\code{parent_company_cik}}{character \link{optional}}

\item{\code{parent_company_state}}{character \link{optional}}

\item{\code{parent_company_SIC_code}}{integer \link{optional}}

\item{\code{purchase_price}}{\link{DimensionalDataPoint} \link{optional}}

\item{\code{trace_link}}{character \link{optional}}

\item{\code{intangible_categories}}{Defined by a Calcbench client named list(\link{IntangibleAssetsCategory}) \link{optional}}

\item{\code{standardized_PPA_points}}{All of the BusinessCombinations... points named list(\link{MappedDataPoint}) \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PurchasePriceAllocation-new}{\code{PurchasePriceAllocation$new()}}
\item \href{#method-PurchasePriceAllocation-toJSON}{\code{PurchasePriceAllocation$toJSON()}}
\item \href{#method-PurchasePriceAllocation-fromJSON}{\code{PurchasePriceAllocation$fromJSON()}}
\item \href{#method-PurchasePriceAllocation-toJSONString}{\code{PurchasePriceAllocation$toJSONString()}}
\item \href{#method-PurchasePriceAllocation-fromJSONString}{\code{PurchasePriceAllocation$fromJSONString()}}
\item \href{#method-PurchasePriceAllocation-validateJSON}{\code{PurchasePriceAllocation$validateJSON()}}
\item \href{#method-PurchasePriceAllocation-toString}{\code{PurchasePriceAllocation$toString()}}
\item \href{#method-PurchasePriceAllocation-isValid}{\code{PurchasePriceAllocation$isValid()}}
\item \href{#method-PurchasePriceAllocation-getInvalidFields}{\code{PurchasePriceAllocation$getInvalidFields()}}
\item \href{#method-PurchasePriceAllocation-print}{\code{PurchasePriceAllocation$print()}}
\item \href{#method-PurchasePriceAllocation-clone}{\code{PurchasePriceAllocation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PurchasePriceAllocation-new"></a>}}
\if{latex}{\out{\hypertarget{method-PurchasePriceAllocation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PurchasePriceAllocation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PurchasePriceAllocation$new(
  acquisition_date = NULL,
  date_reported = NULL,
  date_originally_reported = NULL,
  target = NULL,
  enterprise_value = NULL,
  parent_company = NULL,
  parent_company_ticker = NULL,
  parent_company_cik = NULL,
  parent_company_state = NULL,
  parent_company_SIC_code = NULL,
  purchase_price = NULL,
  trace_link = NULL,
  intangible_categories = NULL,
  standardized_PPA_points = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{acquisition_date}}{acquisition_date}

\item{\code{date_reported}}{date_reported}

\item{\code{date_originally_reported}}{date_originally_reported}

\item{\code{target}}{The company being acquired}

\item{\code{enterprise_value}}{Purchase price + Debt - Cash}

\item{\code{parent_company}}{The acquiring company}

\item{\code{parent_company_ticker}}{parent_company_ticker}

\item{\code{parent_company_cik}}{parent_company_cik}

\item{\code{parent_company_state}}{parent_company_state}

\item{\code{parent_company_SIC_code}}{parent_company_SIC_code}

\item{\code{purchase_price}}{purchase_price}

\item{\code{trace_link}}{trace_link}

\item{\code{intangible_categories}}{Defined by a Calcbench client}

\item{\code{standardized_PPA_points}}{All of the BusinessCombinations... points}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PurchasePriceAllocation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PurchasePriceAllocation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PurchasePriceAllocation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PurchasePriceAllocation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PurchasePriceAllocation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PurchasePriceAllocation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PurchasePriceAllocation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PurchasePriceAllocation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PurchasePriceAllocation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PurchasePriceAllocation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PurchasePriceAllocation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PurchasePriceAllocation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PurchasePriceAllocation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PurchasePriceAllocation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PurchasePriceAllocation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PurchasePriceAllocation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PurchasePriceAllocation$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PurchasePriceAllocation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PurchasePriceAllocation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PurchasePriceAllocation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PurchasePriceAllocation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PurchasePriceAllocation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PurchasePriceAllocation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PurchasePriceAllocation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PurchasePriceAllocation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PurchasePriceAllocation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PurchasePriceAllocation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PurchasePriceAllocation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PurchasePriceAllocation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PurchasePriceAllocation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PurchasePriceAllocation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PurchasePriceAllocation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PurchasePriceAllocation-print"></a>}}
\if{latex}{\out{\hypertarget{method-PurchasePriceAllocation-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PurchasePriceAllocation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PurchasePriceAllocation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PurchasePriceAllocation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PurchasePriceAllocation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
