% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw13_f_data_search_params.R
\docType{class}
\name{Raw13FDataSearchParams}
\alias{Raw13FDataSearchParams}
\title{Raw13FDataSearchParams}
\format{
An \code{R6Class} generator object
}
\description{
Raw13FDataSearchParams Class

Raw13FDataSearchParams Class
}
\details{
Create a new Raw13FDataSearchParams
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{companiesParameters}}{\link{CompaniesParameters} \link{optional}}

\item{\code{pageParameters}}{\link{Raw13FQueryParams} \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Raw13FDataSearchParams-new}{\code{Raw13FDataSearchParams$new()}}
\item \href{#method-Raw13FDataSearchParams-toJSON}{\code{Raw13FDataSearchParams$toJSON()}}
\item \href{#method-Raw13FDataSearchParams-fromJSON}{\code{Raw13FDataSearchParams$fromJSON()}}
\item \href{#method-Raw13FDataSearchParams-toJSONString}{\code{Raw13FDataSearchParams$toJSONString()}}
\item \href{#method-Raw13FDataSearchParams-fromJSONString}{\code{Raw13FDataSearchParams$fromJSONString()}}
\item \href{#method-Raw13FDataSearchParams-validateJSON}{\code{Raw13FDataSearchParams$validateJSON()}}
\item \href{#method-Raw13FDataSearchParams-toString}{\code{Raw13FDataSearchParams$toString()}}
\item \href{#method-Raw13FDataSearchParams-isValid}{\code{Raw13FDataSearchParams$isValid()}}
\item \href{#method-Raw13FDataSearchParams-getInvalidFields}{\code{Raw13FDataSearchParams$getInvalidFields()}}
\item \href{#method-Raw13FDataSearchParams-print}{\code{Raw13FDataSearchParams$print()}}
\item \href{#method-Raw13FDataSearchParams-clone}{\code{Raw13FDataSearchParams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FDataSearchParams-new"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FDataSearchParams-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Raw13FDataSearchParams class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FDataSearchParams$new(
  companiesParameters = NULL,
  pageParameters = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{companiesParameters}}{companiesParameters}

\item{\code{pageParameters}}{pageParameters}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FDataSearchParams-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FDataSearchParams-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FDataSearchParams$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Raw13FDataSearchParams in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FDataSearchParams-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FDataSearchParams-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Raw13FDataSearchParams
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FDataSearchParams$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Raw13FDataSearchParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FDataSearchParams-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FDataSearchParams-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FDataSearchParams$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Raw13FDataSearchParams in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FDataSearchParams-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FDataSearchParams-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Raw13FDataSearchParams
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FDataSearchParams$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Raw13FDataSearchParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FDataSearchParams-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FDataSearchParams-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Raw13FDataSearchParams and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FDataSearchParams$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FDataSearchParams-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FDataSearchParams-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FDataSearchParams$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Raw13FDataSearchParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FDataSearchParams-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FDataSearchParams-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FDataSearchParams$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FDataSearchParams-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FDataSearchParams-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FDataSearchParams$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FDataSearchParams-print"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FDataSearchParams-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FDataSearchParams$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FDataSearchParams-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FDataSearchParams-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FDataSearchParams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
