% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw13_f_query_params.R
\docType{class}
\name{Raw13FQueryParams}
\alias{Raw13FQueryParams}
\title{Raw13FQueryParams}
\format{
An \code{R6Class} generator object
}
\description{
Raw13FQueryParams Class

Raw13FQueryParams Class
}
\details{
Create a new Raw13FQueryParams
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Clauses}}{list(\link{Form13FQueryPiece}) \link{optional}}

\item{\code{QueryID}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Raw13FQueryParams-new}{\code{Raw13FQueryParams$new()}}
\item \href{#method-Raw13FQueryParams-toJSON}{\code{Raw13FQueryParams$toJSON()}}
\item \href{#method-Raw13FQueryParams-fromJSON}{\code{Raw13FQueryParams$fromJSON()}}
\item \href{#method-Raw13FQueryParams-toJSONString}{\code{Raw13FQueryParams$toJSONString()}}
\item \href{#method-Raw13FQueryParams-fromJSONString}{\code{Raw13FQueryParams$fromJSONString()}}
\item \href{#method-Raw13FQueryParams-validateJSON}{\code{Raw13FQueryParams$validateJSON()}}
\item \href{#method-Raw13FQueryParams-toString}{\code{Raw13FQueryParams$toString()}}
\item \href{#method-Raw13FQueryParams-isValid}{\code{Raw13FQueryParams$isValid()}}
\item \href{#method-Raw13FQueryParams-getInvalidFields}{\code{Raw13FQueryParams$getInvalidFields()}}
\item \href{#method-Raw13FQueryParams-print}{\code{Raw13FQueryParams$print()}}
\item \href{#method-Raw13FQueryParams-clone}{\code{Raw13FQueryParams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FQueryParams-new"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FQueryParams-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Raw13FQueryParams class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FQueryParams$new(Clauses = NULL, QueryID = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Clauses}}{Clauses}

\item{\code{QueryID}}{QueryID}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FQueryParams-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FQueryParams-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FQueryParams$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Raw13FQueryParams in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FQueryParams-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FQueryParams-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Raw13FQueryParams
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FQueryParams$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Raw13FQueryParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FQueryParams-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FQueryParams-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FQueryParams$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Raw13FQueryParams in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FQueryParams-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FQueryParams-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Raw13FQueryParams
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FQueryParams$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Raw13FQueryParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FQueryParams-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FQueryParams-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Raw13FQueryParams and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FQueryParams$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FQueryParams-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FQueryParams-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FQueryParams$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Raw13FQueryParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FQueryParams-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FQueryParams-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FQueryParams$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FQueryParams-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FQueryParams-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FQueryParams$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FQueryParams-print"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FQueryParams-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FQueryParams$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Raw13FQueryParams-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Raw13FQueryParams-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Raw13FQueryParams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
