% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_non_xbrl_data_search_params.R
\docType{class}
\name{RawNonXBRLDataSearchParams}
\alias{RawNonXBRLDataSearchParams}
\title{RawNonXBRLDataSearchParams}
\format{
An \code{R6Class} generator object
}
\description{
RawNonXBRLDataSearchParams Class

RawNonXBRLDataSearchParams Class
}
\details{
Create a new RawNonXBRLDataSearchParams
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{companiesParameters}}{\link{CompaniesParameters} \link{optional}}

\item{\code{pageParameters}}{\link{RawNonXBRLQueryParams} \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RawNonXBRLDataSearchParams-new}{\code{RawNonXBRLDataSearchParams$new()}}
\item \href{#method-RawNonXBRLDataSearchParams-toJSON}{\code{RawNonXBRLDataSearchParams$toJSON()}}
\item \href{#method-RawNonXBRLDataSearchParams-fromJSON}{\code{RawNonXBRLDataSearchParams$fromJSON()}}
\item \href{#method-RawNonXBRLDataSearchParams-toJSONString}{\code{RawNonXBRLDataSearchParams$toJSONString()}}
\item \href{#method-RawNonXBRLDataSearchParams-fromJSONString}{\code{RawNonXBRLDataSearchParams$fromJSONString()}}
\item \href{#method-RawNonXBRLDataSearchParams-validateJSON}{\code{RawNonXBRLDataSearchParams$validateJSON()}}
\item \href{#method-RawNonXBRLDataSearchParams-toString}{\code{RawNonXBRLDataSearchParams$toString()}}
\item \href{#method-RawNonXBRLDataSearchParams-isValid}{\code{RawNonXBRLDataSearchParams$isValid()}}
\item \href{#method-RawNonXBRLDataSearchParams-getInvalidFields}{\code{RawNonXBRLDataSearchParams$getInvalidFields()}}
\item \href{#method-RawNonXBRLDataSearchParams-print}{\code{RawNonXBRLDataSearchParams$print()}}
\item \href{#method-RawNonXBRLDataSearchParams-clone}{\code{RawNonXBRLDataSearchParams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLDataSearchParams-new"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLDataSearchParams-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RawNonXBRLDataSearchParams class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLDataSearchParams$new(
  companiesParameters = NULL,
  pageParameters = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{companiesParameters}}{companiesParameters}

\item{\code{pageParameters}}{pageParameters}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLDataSearchParams-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLDataSearchParams-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLDataSearchParams$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RawNonXBRLDataSearchParams in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLDataSearchParams-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLDataSearchParams-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of RawNonXBRLDataSearchParams
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLDataSearchParams$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RawNonXBRLDataSearchParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLDataSearchParams-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLDataSearchParams-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLDataSearchParams$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RawNonXBRLDataSearchParams in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLDataSearchParams-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLDataSearchParams-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of RawNonXBRLDataSearchParams
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLDataSearchParams$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RawNonXBRLDataSearchParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLDataSearchParams-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLDataSearchParams-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to RawNonXBRLDataSearchParams and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLDataSearchParams$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLDataSearchParams-toString"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLDataSearchParams-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLDataSearchParams$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of RawNonXBRLDataSearchParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLDataSearchParams-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLDataSearchParams-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLDataSearchParams$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLDataSearchParams-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLDataSearchParams-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLDataSearchParams$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLDataSearchParams-print"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLDataSearchParams-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLDataSearchParams$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLDataSearchParams-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLDataSearchParams-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLDataSearchParams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
