% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_non_xbrl_query_params.R
\docType{class}
\name{RawNonXBRLQueryParams}
\alias{RawNonXBRLQueryParams}
\title{RawNonXBRLQueryParams}
\format{
An \code{R6Class} generator object
}
\description{
RawNonXBRLQueryParams Class

RawNonXBRLQueryParams Class
}
\details{
Create a new RawNonXBRLQueryParams
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Clauses}}{list(\link{NONxbrlQueryPiece}) \link{optional}}

\item{\code{QueryID}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RawNonXBRLQueryParams-new}{\code{RawNonXBRLQueryParams$new()}}
\item \href{#method-RawNonXBRLQueryParams-toJSON}{\code{RawNonXBRLQueryParams$toJSON()}}
\item \href{#method-RawNonXBRLQueryParams-fromJSON}{\code{RawNonXBRLQueryParams$fromJSON()}}
\item \href{#method-RawNonXBRLQueryParams-toJSONString}{\code{RawNonXBRLQueryParams$toJSONString()}}
\item \href{#method-RawNonXBRLQueryParams-fromJSONString}{\code{RawNonXBRLQueryParams$fromJSONString()}}
\item \href{#method-RawNonXBRLQueryParams-validateJSON}{\code{RawNonXBRLQueryParams$validateJSON()}}
\item \href{#method-RawNonXBRLQueryParams-toString}{\code{RawNonXBRLQueryParams$toString()}}
\item \href{#method-RawNonXBRLQueryParams-isValid}{\code{RawNonXBRLQueryParams$isValid()}}
\item \href{#method-RawNonXBRLQueryParams-getInvalidFields}{\code{RawNonXBRLQueryParams$getInvalidFields()}}
\item \href{#method-RawNonXBRLQueryParams-print}{\code{RawNonXBRLQueryParams$print()}}
\item \href{#method-RawNonXBRLQueryParams-clone}{\code{RawNonXBRLQueryParams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLQueryParams-new"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLQueryParams-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RawNonXBRLQueryParams class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLQueryParams$new(Clauses = NULL, QueryID = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Clauses}}{Clauses}

\item{\code{QueryID}}{QueryID}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLQueryParams-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLQueryParams-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLQueryParams$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RawNonXBRLQueryParams in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLQueryParams-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLQueryParams-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of RawNonXBRLQueryParams
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLQueryParams$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RawNonXBRLQueryParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLQueryParams-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLQueryParams-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLQueryParams$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RawNonXBRLQueryParams in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLQueryParams-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLQueryParams-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of RawNonXBRLQueryParams
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLQueryParams$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RawNonXBRLQueryParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLQueryParams-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLQueryParams-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to RawNonXBRLQueryParams and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLQueryParams$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLQueryParams-toString"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLQueryParams-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLQueryParams$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of RawNonXBRLQueryParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLQueryParams-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLQueryParams-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLQueryParams$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLQueryParams-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLQueryParams-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLQueryParams$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLQueryParams-print"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLQueryParams-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLQueryParams$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawNonXBRLQueryParams-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RawNonXBRLQueryParams-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawNonXBRLQueryParams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
