% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_xbrl_api.R
\docType{class}
\name{RawXBRLApi}
\alias{RawXBRLApi}
\title{RawXBRL operations}
\format{
An \code{R6Class} generator object
}
\description{
RawXBRLApi
}
\details{
Calcbench API

No description provided (generated by Openapi Generator https://github.com/calcbenchRtools/calcbenchR-generator)

The version of the OpenAPI document: v1
Generated by: https://calcbenchR-generator.tech
}
\section{Methods}{

\describe{
\strong{ RawXBRLXBRLTagValues } \emph{ Get data for individual tags on the multi-company page. }
See <a href=\"https://www.calcbench.com/multi?pg_classificationMethod=index&pg_title=Dow\\%20Jones\\%20Industrial\\%20Average&pg_sectorCode=DJIA&mc_metrics=\\%7B\\%22metric\\%22:\\%22Revenues\\%22,\\%22dataSet\\%22:2,\\%22format_type\\%22:\\%22currency\\%22\\%7D\">Example\if{html}{\out{</a>}}

\itemize{
\item \emph{ @param } search_parameters \link{XBRLTagSearchParameters}

\item status code : 200 | OK

\item return type : array\link{object}
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  RawXBRLXBRLTagValues  ####################

library(calcbenchR)
var_search_parameters <- XBRLTagSearchParameters$new(XBRLTagQueryParameters$new(c("tags_example"), c("metrics_example")), CompaniesParameters$new(c("companyIdentifiers_example"), "entireUniverse_example", c(123)), PeriodParameters$new(123, "unset", 123, "unset", "periodType_example", "useFiscalPeriod_example", DateRange$new("startDate_example", "endDate_example"), "allHistory_example", "updateDate_example", "updatedFrom_example", "asOriginallyReported_example", 123, 123), MetricSearchCriteria$new(c("metrics_example"), c("company_identifiers_example"), 123, 123, 123, 123, "fiscal_period_example", 123, "calendar_period_example", 123, "include_trace_example", "entire_universe_example", 123, "use_fiscal_periods_example", "as_originally_reported_example", "all_history_example", "PointInTime_example", "updateDate_example", "unset", "Annual_example", "Quarterly_example", c("unset"), c(key = "inner_example"))) # XBRLTagSearchParameters | 

#Get data for individual tags on the multi-company page.
api_instance <- RawXBRLApi$new()

# to save the result into a file, simply add the optional `data_file` parameter, e.g.
# result <- api_instance$RawXBRLXBRLTagValues(var_search_parametersdata_file = "result.txt")
result <- api_instance$RawXBRLXBRLTagValues(var_search_parameters)
dput(result)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RawXBRLApi-new}{\code{RawXBRLApi$new()}}
\item \href{#method-RawXBRLApi-RawXBRLXBRLTagValues}{\code{RawXBRLApi$RawXBRLXBRLTagValues()}}
\item \href{#method-RawXBRLApi-RawXBRLXBRLTagValuesWithHttpInfo}{\code{RawXBRLApi$RawXBRLXBRLTagValuesWithHttpInfo()}}
\item \href{#method-RawXBRLApi-clone}{\code{RawXBRLApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RawXBRLApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLApi-RawXBRLXBRLTagValues"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLApi-RawXBRLXBRLTagValues}{}}}
\subsection{Method \code{RawXBRLXBRLTagValues()}}{
Get data for individual tags on the multi-company page.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLApi$RawXBRLXBRLTagValues(search_parameters, data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_parameters}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array\link{object}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLApi-RawXBRLXBRLTagValuesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLApi-RawXBRLXBRLTagValuesWithHttpInfo}{}}}
\subsection{Method \code{RawXBRLXBRLTagValuesWithHttpInfo()}}{
Get data for individual tags on the multi-company page.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLApi$RawXBRLXBRLTagValuesWithHttpInfo(
  search_parameters,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_parameters}}{}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (array\link{object}) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
