% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_xbrl_data_search_params.R
\docType{class}
\name{RawXBRLDataSearchParams}
\alias{RawXBRLDataSearchParams}
\title{RawXBRLDataSearchParams}
\format{
An \code{R6Class} generator object
}
\description{
RawXBRLDataSearchParams Class

RawXBRLDataSearchParams Class
}
\details{
Create a new RawXBRLDataSearchParams
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pageParameters}}{\link{RawXBRLQueryParams} \link{optional}}

\item{\code{companiesParameters}}{\link{CompaniesParameters} \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RawXBRLDataSearchParams-new}{\code{RawXBRLDataSearchParams$new()}}
\item \href{#method-RawXBRLDataSearchParams-toJSON}{\code{RawXBRLDataSearchParams$toJSON()}}
\item \href{#method-RawXBRLDataSearchParams-fromJSON}{\code{RawXBRLDataSearchParams$fromJSON()}}
\item \href{#method-RawXBRLDataSearchParams-toJSONString}{\code{RawXBRLDataSearchParams$toJSONString()}}
\item \href{#method-RawXBRLDataSearchParams-fromJSONString}{\code{RawXBRLDataSearchParams$fromJSONString()}}
\item \href{#method-RawXBRLDataSearchParams-validateJSON}{\code{RawXBRLDataSearchParams$validateJSON()}}
\item \href{#method-RawXBRLDataSearchParams-toString}{\code{RawXBRLDataSearchParams$toString()}}
\item \href{#method-RawXBRLDataSearchParams-isValid}{\code{RawXBRLDataSearchParams$isValid()}}
\item \href{#method-RawXBRLDataSearchParams-getInvalidFields}{\code{RawXBRLDataSearchParams$getInvalidFields()}}
\item \href{#method-RawXBRLDataSearchParams-print}{\code{RawXBRLDataSearchParams$print()}}
\item \href{#method-RawXBRLDataSearchParams-clone}{\code{RawXBRLDataSearchParams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLDataSearchParams-new"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLDataSearchParams-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RawXBRLDataSearchParams class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLDataSearchParams$new(
  pageParameters = NULL,
  companiesParameters = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pageParameters}}{pageParameters}

\item{\code{companiesParameters}}{companiesParameters}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLDataSearchParams-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLDataSearchParams-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLDataSearchParams$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RawXBRLDataSearchParams in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLDataSearchParams-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLDataSearchParams-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of RawXBRLDataSearchParams
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLDataSearchParams$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RawXBRLDataSearchParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLDataSearchParams-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLDataSearchParams-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLDataSearchParams$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RawXBRLDataSearchParams in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLDataSearchParams-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLDataSearchParams-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of RawXBRLDataSearchParams
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLDataSearchParams$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RawXBRLDataSearchParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLDataSearchParams-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLDataSearchParams-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to RawXBRLDataSearchParams and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLDataSearchParams$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLDataSearchParams-toString"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLDataSearchParams-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLDataSearchParams$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of RawXBRLDataSearchParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLDataSearchParams-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLDataSearchParams-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLDataSearchParams$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLDataSearchParams-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLDataSearchParams-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLDataSearchParams$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLDataSearchParams-print"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLDataSearchParams-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLDataSearchParams$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLDataSearchParams-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLDataSearchParams-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLDataSearchParams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
