% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_xbrl_query_params.R
\docType{class}
\name{RawXBRLQueryParams}
\alias{RawXBRLQueryParams}
\title{RawXBRLQueryParams}
\format{
An \code{R6Class} generator object
}
\description{
RawXBRLQueryParams Class

RawXBRLQueryParams Class
}
\details{
Create a new RawXBRLQueryParams
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Clauses}}{list(\link{XbrlQueryPiece}) \link{optional}}

\item{\code{QueryID}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RawXBRLQueryParams-new}{\code{RawXBRLQueryParams$new()}}
\item \href{#method-RawXBRLQueryParams-toJSON}{\code{RawXBRLQueryParams$toJSON()}}
\item \href{#method-RawXBRLQueryParams-fromJSON}{\code{RawXBRLQueryParams$fromJSON()}}
\item \href{#method-RawXBRLQueryParams-toJSONString}{\code{RawXBRLQueryParams$toJSONString()}}
\item \href{#method-RawXBRLQueryParams-fromJSONString}{\code{RawXBRLQueryParams$fromJSONString()}}
\item \href{#method-RawXBRLQueryParams-validateJSON}{\code{RawXBRLQueryParams$validateJSON()}}
\item \href{#method-RawXBRLQueryParams-toString}{\code{RawXBRLQueryParams$toString()}}
\item \href{#method-RawXBRLQueryParams-isValid}{\code{RawXBRLQueryParams$isValid()}}
\item \href{#method-RawXBRLQueryParams-getInvalidFields}{\code{RawXBRLQueryParams$getInvalidFields()}}
\item \href{#method-RawXBRLQueryParams-print}{\code{RawXBRLQueryParams$print()}}
\item \href{#method-RawXBRLQueryParams-clone}{\code{RawXBRLQueryParams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLQueryParams-new"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLQueryParams-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RawXBRLQueryParams class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLQueryParams$new(Clauses = NULL, QueryID = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Clauses}}{Clauses}

\item{\code{QueryID}}{QueryID}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLQueryParams-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLQueryParams-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLQueryParams$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RawXBRLQueryParams in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLQueryParams-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLQueryParams-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of RawXBRLQueryParams
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLQueryParams$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RawXBRLQueryParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLQueryParams-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLQueryParams-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLQueryParams$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RawXBRLQueryParams in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLQueryParams-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLQueryParams-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of RawXBRLQueryParams
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLQueryParams$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RawXBRLQueryParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLQueryParams-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLQueryParams-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to RawXBRLQueryParams and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLQueryParams$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLQueryParams-toString"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLQueryParams-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLQueryParams$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of RawXBRLQueryParams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLQueryParams-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLQueryParams-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLQueryParams$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLQueryParams-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLQueryParams-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLQueryParams$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLQueryParams-print"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLQueryParams-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLQueryParams$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RawXBRLQueryParams-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RawXBRLQueryParams-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawXBRLQueryParams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
