% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_footnote_metric_definition.R
\docType{class}
\name{StandardizedFootnoteMetricDefinition}
\alias{StandardizedFootnoteMetricDefinition}
\title{StandardizedFootnoteMetricDefinition}
\format{
An \code{R6Class} generator object
}
\description{
StandardizedFootnoteMetricDefinition Class

StandardizedFootnoteMetricDefinition Class
}
\details{
Create a new StandardizedFootnoteMetricDefinition
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{related_footnotes}}{list(character) \link{optional}}

\item{\code{indent_level}}{integer \link{optional}}

\item{\code{metric}}{character \link{optional}}

\item{\code{name}}{character \link{optional}}

\item{\code{period_type}}{character \link{optional}}

\item{\code{section}}{character \link{optional}}

\item{\code{class}}{character \link{optional}}

\item{\code{equation}}{character \link{optional}}

\item{\code{format_type}}{character \link{optional}}

\item{\code{is_face}}{character \link{optional}}

\item{\code{is_footnote}}{character \link{optional}}

\item{\code{is_ratio}}{character \link{optional}}

\item{\code{description}}{character \link{optional}}

\item{\code{UnitOfMeasure}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StandardizedFootnoteMetricDefinition-new}{\code{StandardizedFootnoteMetricDefinition$new()}}
\item \href{#method-StandardizedFootnoteMetricDefinition-toJSON}{\code{StandardizedFootnoteMetricDefinition$toJSON()}}
\item \href{#method-StandardizedFootnoteMetricDefinition-fromJSON}{\code{StandardizedFootnoteMetricDefinition$fromJSON()}}
\item \href{#method-StandardizedFootnoteMetricDefinition-toJSONString}{\code{StandardizedFootnoteMetricDefinition$toJSONString()}}
\item \href{#method-StandardizedFootnoteMetricDefinition-fromJSONString}{\code{StandardizedFootnoteMetricDefinition$fromJSONString()}}
\item \href{#method-StandardizedFootnoteMetricDefinition-validateJSON}{\code{StandardizedFootnoteMetricDefinition$validateJSON()}}
\item \href{#method-StandardizedFootnoteMetricDefinition-toString}{\code{StandardizedFootnoteMetricDefinition$toString()}}
\item \href{#method-StandardizedFootnoteMetricDefinition-isValid}{\code{StandardizedFootnoteMetricDefinition$isValid()}}
\item \href{#method-StandardizedFootnoteMetricDefinition-getInvalidFields}{\code{StandardizedFootnoteMetricDefinition$getInvalidFields()}}
\item \href{#method-StandardizedFootnoteMetricDefinition-print}{\code{StandardizedFootnoteMetricDefinition$print()}}
\item \href{#method-StandardizedFootnoteMetricDefinition-clone}{\code{StandardizedFootnoteMetricDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedFootnoteMetricDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedFootnoteMetricDefinition-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new StandardizedFootnoteMetricDefinition class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedFootnoteMetricDefinition$new(
  related_footnotes = NULL,
  indent_level = NULL,
  metric = NULL,
  name = NULL,
  period_type = NULL,
  section = NULL,
  class = NULL,
  equation = NULL,
  format_type = NULL,
  is_face = NULL,
  is_footnote = NULL,
  is_ratio = NULL,
  description = NULL,
  UnitOfMeasure = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{related_footnotes}}{related_footnotes}

\item{\code{indent_level}}{indent_level}

\item{\code{metric}}{metric}

\item{\code{name}}{name}

\item{\code{period_type}}{period_type}

\item{\code{section}}{section}

\item{\code{class}}{class}

\item{\code{equation}}{equation}

\item{\code{format_type}}{format_type}

\item{\code{is_face}}{is_face}

\item{\code{is_footnote}}{is_footnote}

\item{\code{is_ratio}}{is_ratio}

\item{\code{description}}{description}

\item{\code{UnitOfMeasure}}{UnitOfMeasure}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedFootnoteMetricDefinition-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedFootnoteMetricDefinition-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedFootnoteMetricDefinition$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StandardizedFootnoteMetricDefinition in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedFootnoteMetricDefinition-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedFootnoteMetricDefinition-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of StandardizedFootnoteMetricDefinition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedFootnoteMetricDefinition$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StandardizedFootnoteMetricDefinition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedFootnoteMetricDefinition-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedFootnoteMetricDefinition-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedFootnoteMetricDefinition$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StandardizedFootnoteMetricDefinition in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedFootnoteMetricDefinition-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedFootnoteMetricDefinition-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of StandardizedFootnoteMetricDefinition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedFootnoteMetricDefinition$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StandardizedFootnoteMetricDefinition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedFootnoteMetricDefinition-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedFootnoteMetricDefinition-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to StandardizedFootnoteMetricDefinition and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedFootnoteMetricDefinition$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedFootnoteMetricDefinition-toString"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedFootnoteMetricDefinition-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedFootnoteMetricDefinition$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of StandardizedFootnoteMetricDefinition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedFootnoteMetricDefinition-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedFootnoteMetricDefinition-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedFootnoteMetricDefinition$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedFootnoteMetricDefinition-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedFootnoteMetricDefinition-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedFootnoteMetricDefinition$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedFootnoteMetricDefinition-print"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedFootnoteMetricDefinition-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedFootnoteMetricDefinition$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedFootnoteMetricDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedFootnoteMetricDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedFootnoteMetricDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
