% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_metric_definition.R
\docType{class}
\name{StandardizedMetricDefinition}
\alias{StandardizedMetricDefinition}
\title{StandardizedMetricDefinition}
\format{
An \code{R6Class} generator object
}
\description{
StandardizedMetricDefinition Class

StandardizedMetricDefinition Class
}
\details{
Create a new StandardizedMetricDefinition
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metric}}{character \link{optional}}

\item{\code{name}}{character \link{optional}}

\item{\code{period_type}}{character \link{optional}}

\item{\code{section}}{character \link{optional}}

\item{\code{class}}{character \link{optional}}

\item{\code{equation}}{character \link{optional}}

\item{\code{format_type}}{character \link{optional}}

\item{\code{is_face}}{character \link{optional}}

\item{\code{is_footnote}}{character \link{optional}}

\item{\code{is_ratio}}{character \link{optional}}

\item{\code{description}}{character \link{optional}}

\item{\code{UnitOfMeasure}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StandardizedMetricDefinition-new}{\code{StandardizedMetricDefinition$new()}}
\item \href{#method-StandardizedMetricDefinition-toJSON}{\code{StandardizedMetricDefinition$toJSON()}}
\item \href{#method-StandardizedMetricDefinition-fromJSON}{\code{StandardizedMetricDefinition$fromJSON()}}
\item \href{#method-StandardizedMetricDefinition-toJSONString}{\code{StandardizedMetricDefinition$toJSONString()}}
\item \href{#method-StandardizedMetricDefinition-fromJSONString}{\code{StandardizedMetricDefinition$fromJSONString()}}
\item \href{#method-StandardizedMetricDefinition-validateJSON}{\code{StandardizedMetricDefinition$validateJSON()}}
\item \href{#method-StandardizedMetricDefinition-toString}{\code{StandardizedMetricDefinition$toString()}}
\item \href{#method-StandardizedMetricDefinition-isValid}{\code{StandardizedMetricDefinition$isValid()}}
\item \href{#method-StandardizedMetricDefinition-getInvalidFields}{\code{StandardizedMetricDefinition$getInvalidFields()}}
\item \href{#method-StandardizedMetricDefinition-print}{\code{StandardizedMetricDefinition$print()}}
\item \href{#method-StandardizedMetricDefinition-clone}{\code{StandardizedMetricDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedMetricDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedMetricDefinition-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new StandardizedMetricDefinition class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedMetricDefinition$new(
  metric = NULL,
  name = NULL,
  period_type = NULL,
  section = NULL,
  class = NULL,
  equation = NULL,
  format_type = NULL,
  is_face = NULL,
  is_footnote = NULL,
  is_ratio = NULL,
  description = NULL,
  UnitOfMeasure = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metric}}{metric}

\item{\code{name}}{name}

\item{\code{period_type}}{period_type}

\item{\code{section}}{section}

\item{\code{class}}{class}

\item{\code{equation}}{equation}

\item{\code{format_type}}{format_type}

\item{\code{is_face}}{is_face}

\item{\code{is_footnote}}{is_footnote}

\item{\code{is_ratio}}{is_ratio}

\item{\code{description}}{description}

\item{\code{UnitOfMeasure}}{UnitOfMeasure}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedMetricDefinition-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedMetricDefinition-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedMetricDefinition$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StandardizedMetricDefinition in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedMetricDefinition-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedMetricDefinition-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of StandardizedMetricDefinition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedMetricDefinition$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StandardizedMetricDefinition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedMetricDefinition-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedMetricDefinition-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedMetricDefinition$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StandardizedMetricDefinition in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedMetricDefinition-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedMetricDefinition-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of StandardizedMetricDefinition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedMetricDefinition$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StandardizedMetricDefinition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedMetricDefinition-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedMetricDefinition-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to StandardizedMetricDefinition and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedMetricDefinition$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedMetricDefinition-toString"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedMetricDefinition-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedMetricDefinition$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of StandardizedMetricDefinition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedMetricDefinition-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedMetricDefinition-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedMetricDefinition$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedMetricDefinition-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedMetricDefinition-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedMetricDefinition$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedMetricDefinition-print"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedMetricDefinition-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedMetricDefinition$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedMetricDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedMetricDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedMetricDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
