% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_query_parameters.R
\docType{class}
\name{StandardizedQueryParameters}
\alias{StandardizedQueryParameters}
\title{StandardizedQueryParameters}
\format{
An \code{R6Class} generator object
}
\description{
StandardizedQueryParameters Class

StandardizedQueryParameters Class
}
\details{
Create a new StandardizedQueryParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pointInTime}}{character \link{optional}}

\item{\code{pointInTimeV2}}{character \link{optional}}

\item{\code{allFootnotes}}{character \link{optional}}

\item{\code{allFace}}{character \link{optional}}

\item{\code{includeXBRL}}{character \link{optional}}

\item{\code{includePreliminary}}{character \link{optional}}

\item{\code{allNonGAAP}}{character \link{optional}}

\item{\code{allMetrics}}{character \link{optional}}

\item{\code{excludeUnconfirmedPreliminary}}{character \link{optional}}

\item{\code{metrics}}{list(character) \link{optional}}

\item{\code{includeTrace}}{character \link{optional}}

\item{\code{excludeErrors}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StandardizedQueryParameters-new}{\code{StandardizedQueryParameters$new()}}
\item \href{#method-StandardizedQueryParameters-toJSON}{\code{StandardizedQueryParameters$toJSON()}}
\item \href{#method-StandardizedQueryParameters-fromJSON}{\code{StandardizedQueryParameters$fromJSON()}}
\item \href{#method-StandardizedQueryParameters-toJSONString}{\code{StandardizedQueryParameters$toJSONString()}}
\item \href{#method-StandardizedQueryParameters-fromJSONString}{\code{StandardizedQueryParameters$fromJSONString()}}
\item \href{#method-StandardizedQueryParameters-validateJSON}{\code{StandardizedQueryParameters$validateJSON()}}
\item \href{#method-StandardizedQueryParameters-toString}{\code{StandardizedQueryParameters$toString()}}
\item \href{#method-StandardizedQueryParameters-isValid}{\code{StandardizedQueryParameters$isValid()}}
\item \href{#method-StandardizedQueryParameters-getInvalidFields}{\code{StandardizedQueryParameters$getInvalidFields()}}
\item \href{#method-StandardizedQueryParameters-print}{\code{StandardizedQueryParameters$print()}}
\item \href{#method-StandardizedQueryParameters-clone}{\code{StandardizedQueryParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedQueryParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedQueryParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new StandardizedQueryParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedQueryParameters$new(
  pointInTime = NULL,
  pointInTimeV2 = NULL,
  allFootnotes = NULL,
  allFace = NULL,
  includeXBRL = NULL,
  includePreliminary = NULL,
  allNonGAAP = NULL,
  allMetrics = NULL,
  excludeUnconfirmedPreliminary = NULL,
  metrics = NULL,
  includeTrace = NULL,
  excludeErrors = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pointInTime}}{pointInTime}

\item{\code{pointInTimeV2}}{pointInTimeV2}

\item{\code{allFootnotes}}{allFootnotes}

\item{\code{allFace}}{allFace}

\item{\code{includeXBRL}}{includeXBRL}

\item{\code{includePreliminary}}{includePreliminary}

\item{\code{allNonGAAP}}{allNonGAAP}

\item{\code{allMetrics}}{allMetrics}

\item{\code{excludeUnconfirmedPreliminary}}{excludeUnconfirmedPreliminary}

\item{\code{metrics}}{metrics}

\item{\code{includeTrace}}{includeTrace}

\item{\code{excludeErrors}}{excludeErrors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedQueryParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedQueryParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedQueryParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StandardizedQueryParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedQueryParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedQueryParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of StandardizedQueryParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedQueryParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StandardizedQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedQueryParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedQueryParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedQueryParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StandardizedQueryParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedQueryParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedQueryParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of StandardizedQueryParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedQueryParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StandardizedQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedQueryParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedQueryParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to StandardizedQueryParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedQueryParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedQueryParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedQueryParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedQueryParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of StandardizedQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedQueryParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedQueryParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedQueryParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedQueryParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedQueryParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedQueryParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedQueryParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedQueryParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedQueryParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardizedQueryParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StandardizedQueryParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardizedQueryParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
