% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements_meta_data.R
\docType{class}
\name{StatementsMetaData}
\alias{StatementsMetaData}
\title{StatementsMetaData}
\format{
An \code{R6Class} generator object
}
\description{
The face statements for a company/period along with some stuff to render the CID page

StatementsMetaData Class
}
\details{
Create a new StatementsMetaData
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{face_statements}}{list(\link{XBRLstatement}) \link{optional}}

\item{\code{context}}{\link{XBRLcontext} \link{optional}}

\item{\code{quickReportsHTML}}{character \link{optional}}

\item{\code{SEClink}}{character \link{optional}}

\item{\code{relatedCompanies}}{list(\link{XBRLcompany}) \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StatementsMetaData-new}{\code{StatementsMetaData$new()}}
\item \href{#method-StatementsMetaData-toJSON}{\code{StatementsMetaData$toJSON()}}
\item \href{#method-StatementsMetaData-fromJSON}{\code{StatementsMetaData$fromJSON()}}
\item \href{#method-StatementsMetaData-toJSONString}{\code{StatementsMetaData$toJSONString()}}
\item \href{#method-StatementsMetaData-fromJSONString}{\code{StatementsMetaData$fromJSONString()}}
\item \href{#method-StatementsMetaData-validateJSON}{\code{StatementsMetaData$validateJSON()}}
\item \href{#method-StatementsMetaData-toString}{\code{StatementsMetaData$toString()}}
\item \href{#method-StatementsMetaData-isValid}{\code{StatementsMetaData$isValid()}}
\item \href{#method-StatementsMetaData-getInvalidFields}{\code{StatementsMetaData$getInvalidFields()}}
\item \href{#method-StatementsMetaData-print}{\code{StatementsMetaData$print()}}
\item \href{#method-StatementsMetaData-clone}{\code{StatementsMetaData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatementsMetaData-new"></a>}}
\if{latex}{\out{\hypertarget{method-StatementsMetaData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new StatementsMetaData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatementsMetaData$new(
  face_statements = NULL,
  context = NULL,
  quickReportsHTML = NULL,
  SEClink = NULL,
  relatedCompanies = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{face_statements}}{face_statements}

\item{\code{context}}{context}

\item{\code{quickReportsHTML}}{quickReportsHTML}

\item{\code{SEClink}}{SEClink}

\item{\code{relatedCompanies}}{relatedCompanies}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatementsMetaData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StatementsMetaData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatementsMetaData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StatementsMetaData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatementsMetaData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StatementsMetaData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of StatementsMetaData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatementsMetaData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StatementsMetaData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatementsMetaData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StatementsMetaData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatementsMetaData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StatementsMetaData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatementsMetaData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StatementsMetaData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of StatementsMetaData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatementsMetaData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StatementsMetaData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatementsMetaData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StatementsMetaData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to StatementsMetaData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatementsMetaData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatementsMetaData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-StatementsMetaData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatementsMetaData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of StatementsMetaData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatementsMetaData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-StatementsMetaData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatementsMetaData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatementsMetaData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-StatementsMetaData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatementsMetaData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatementsMetaData-print"></a>}}
\if{latex}{\out{\hypertarget{method-StatementsMetaData-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatementsMetaData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatementsMetaData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatementsMetaData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatementsMetaData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
