% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_disclosure.R
\docType{class}
\name{XBRLDisclosure}
\alias{XBRLDisclosure}
\title{XBRLDisclosure}
\format{
An \code{R6Class} generator object
}
\description{
XBRLDisclosure Class

XBRLDisclosure Class
}
\details{
Create a new XBRLDisclosure
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entity_id}}{object \link{optional}}

\item{\code{entity_name}}{object \link{optional}}

\item{\code{document_type}}{character \link{optional}}

\item{\code{end_date}}{character \link{optional}}

\item{\code{filing_date}}{character \link{optional}}

\item{\code{sec_html_url}}{character \link{optional}}

\item{\code{sec_accession_number}}{character \link{optional}}

\item{\code{accession_id}}{integer \link{optional}}

\item{\code{label}}{character \link{optional}}

\item{\code{fact_id}}{integer \link{optional}}

\item{\code{disclosure_type}}{character \link{optional}}

\item{\code{is_detail}}{character \link{optional}}

\item{\code{fiscal_period}}{character \link{optional}}

\item{\code{fiscal_year}}{integer \link{optional}}

\item{\code{last_in_group}}{character \link{optional}}

\item{\code{networkID}}{integer \link{optional}}

\item{\code{ticker}}{character \link{optional}}

\item{\code{table_list}}{list(object) \link{optional}}

\item{\code{local_name}}{character \link{optional}}

\item{\code{CIK}}{character \link{optional}}

\item{\code{facts}}{object \link{optional}}

\item{\code{blobs}}{list(character) \link{optional}}

\item{\code{date_reported}}{character \link{optional}}

\item{\code{display_filing_date}}{character \link{optional}}

\item{\code{display_end_date}}{character \link{optional}}

\item{\code{CalendarPeriod}}{character \link{optional}}

\item{\code{CalendarYear}}{integer \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XBRLDisclosure-new}{\code{XBRLDisclosure$new()}}
\item \href{#method-XBRLDisclosure-toJSON}{\code{XBRLDisclosure$toJSON()}}
\item \href{#method-XBRLDisclosure-fromJSON}{\code{XBRLDisclosure$fromJSON()}}
\item \href{#method-XBRLDisclosure-toJSONString}{\code{XBRLDisclosure$toJSONString()}}
\item \href{#method-XBRLDisclosure-fromJSONString}{\code{XBRLDisclosure$fromJSONString()}}
\item \href{#method-XBRLDisclosure-validateJSON}{\code{XBRLDisclosure$validateJSON()}}
\item \href{#method-XBRLDisclosure-toString}{\code{XBRLDisclosure$toString()}}
\item \href{#method-XBRLDisclosure-isValid}{\code{XBRLDisclosure$isValid()}}
\item \href{#method-XBRLDisclosure-getInvalidFields}{\code{XBRLDisclosure$getInvalidFields()}}
\item \href{#method-XBRLDisclosure-print}{\code{XBRLDisclosure$print()}}
\item \href{#method-XBRLDisclosure-clone}{\code{XBRLDisclosure$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLDisclosure-new"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLDisclosure-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new XBRLDisclosure class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLDisclosure$new(
  entity_id = NULL,
  entity_name = NULL,
  document_type = NULL,
  end_date = NULL,
  filing_date = NULL,
  sec_html_url = NULL,
  sec_accession_number = NULL,
  accession_id = NULL,
  label = NULL,
  fact_id = NULL,
  disclosure_type = NULL,
  is_detail = NULL,
  fiscal_period = NULL,
  fiscal_year = NULL,
  last_in_group = NULL,
  networkID = NULL,
  ticker = NULL,
  table_list = NULL,
  local_name = NULL,
  CIK = NULL,
  facts = NULL,
  blobs = NULL,
  date_reported = NULL,
  display_filing_date = NULL,
  display_end_date = NULL,
  CalendarPeriod = NULL,
  CalendarYear = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity_id}}{entity_id}

\item{\code{entity_name}}{entity_name}

\item{\code{document_type}}{document_type}

\item{\code{end_date}}{end_date}

\item{\code{filing_date}}{filing_date}

\item{\code{sec_html_url}}{sec_html_url}

\item{\code{sec_accession_number}}{sec_accession_number}

\item{\code{accession_id}}{accession_id}

\item{\code{label}}{label}

\item{\code{fact_id}}{fact_id}

\item{\code{disclosure_type}}{disclosure_type}

\item{\code{is_detail}}{is_detail}

\item{\code{fiscal_period}}{fiscal_period}

\item{\code{fiscal_year}}{fiscal_year}

\item{\code{last_in_group}}{last_in_group}

\item{\code{networkID}}{networkID}

\item{\code{ticker}}{ticker}

\item{\code{table_list}}{table_list}

\item{\code{local_name}}{local_name}

\item{\code{CIK}}{CIK}

\item{\code{facts}}{facts}

\item{\code{blobs}}{blobs}

\item{\code{date_reported}}{date_reported}

\item{\code{display_filing_date}}{display_filing_date}

\item{\code{display_end_date}}{display_end_date}

\item{\code{CalendarPeriod}}{CalendarPeriod}

\item{\code{CalendarYear}}{CalendarYear}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLDisclosure-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLDisclosure-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLDisclosure$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
XBRLDisclosure in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLDisclosure-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLDisclosure-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of XBRLDisclosure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLDisclosure$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of XBRLDisclosure
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLDisclosure-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLDisclosure-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLDisclosure$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
XBRLDisclosure in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLDisclosure-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLDisclosure-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of XBRLDisclosure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLDisclosure$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of XBRLDisclosure
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLDisclosure-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLDisclosure-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to XBRLDisclosure and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLDisclosure$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLDisclosure-toString"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLDisclosure-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLDisclosure$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of XBRLDisclosure
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLDisclosure-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLDisclosure-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLDisclosure$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLDisclosure-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLDisclosure-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLDisclosure$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLDisclosure-print"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLDisclosure-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLDisclosure$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLDisclosure-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLDisclosure-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLDisclosure$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
