% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_query_piece.R
\docType{class}
\name{XbrlQueryPiece}
\alias{XbrlQueryPiece}
\title{XbrlQueryPiece}
\format{
An \code{R6Class} generator object
}
\description{
XbrlQueryPiece Class

XbrlQueryPiece Class
}
\details{
Create a new XbrlQueryPiece
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parameter}}{character \link{optional}}

\item{\code{Operator}}{character \link{optional}}

\item{\code{value}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XbrlQueryPiece-new}{\code{XbrlQueryPiece$new()}}
\item \href{#method-XbrlQueryPiece-toJSON}{\code{XbrlQueryPiece$toJSON()}}
\item \href{#method-XbrlQueryPiece-fromJSON}{\code{XbrlQueryPiece$fromJSON()}}
\item \href{#method-XbrlQueryPiece-toJSONString}{\code{XbrlQueryPiece$toJSONString()}}
\item \href{#method-XbrlQueryPiece-fromJSONString}{\code{XbrlQueryPiece$fromJSONString()}}
\item \href{#method-XbrlQueryPiece-validateJSON}{\code{XbrlQueryPiece$validateJSON()}}
\item \href{#method-XbrlQueryPiece-toString}{\code{XbrlQueryPiece$toString()}}
\item \href{#method-XbrlQueryPiece-isValid}{\code{XbrlQueryPiece$isValid()}}
\item \href{#method-XbrlQueryPiece-getInvalidFields}{\code{XbrlQueryPiece$getInvalidFields()}}
\item \href{#method-XbrlQueryPiece-print}{\code{XbrlQueryPiece$print()}}
\item \href{#method-XbrlQueryPiece-clone}{\code{XbrlQueryPiece$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XbrlQueryPiece-new"></a>}}
\if{latex}{\out{\hypertarget{method-XbrlQueryPiece-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new XbrlQueryPiece class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XbrlQueryPiece$new(parameter = NULL, Operator = NULL, value = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameter}}{parameter}

\item{\code{Operator}}{Operator}

\item{\code{value}}{value}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XbrlQueryPiece-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XbrlQueryPiece-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XbrlQueryPiece$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
XbrlQueryPiece in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XbrlQueryPiece-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XbrlQueryPiece-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of XbrlQueryPiece
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XbrlQueryPiece$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of XbrlQueryPiece
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XbrlQueryPiece-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-XbrlQueryPiece-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XbrlQueryPiece$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
XbrlQueryPiece in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XbrlQueryPiece-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-XbrlQueryPiece-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of XbrlQueryPiece
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XbrlQueryPiece$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of XbrlQueryPiece
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XbrlQueryPiece-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XbrlQueryPiece-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to XbrlQueryPiece and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XbrlQueryPiece$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XbrlQueryPiece-toString"></a>}}
\if{latex}{\out{\hypertarget{method-XbrlQueryPiece-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XbrlQueryPiece$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of XbrlQueryPiece
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XbrlQueryPiece-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-XbrlQueryPiece-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XbrlQueryPiece$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XbrlQueryPiece-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-XbrlQueryPiece-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XbrlQueryPiece$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XbrlQueryPiece-print"></a>}}
\if{latex}{\out{\hypertarget{method-XbrlQueryPiece-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XbrlQueryPiece$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XbrlQueryPiece-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XbrlQueryPiece-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XbrlQueryPiece$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
