% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbr_lstatement.R
\docType{class}
\name{XBRLstatement}
\alias{XBRLstatement}
\title{XBRLstatement}
\format{
An \code{R6Class} generator object
}
\description{
A \"face\" statement.

XBRLstatement Class
}
\details{
Create a new XBRLstatement
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character \link{optional}}

\item{\code{statementType}}{character \link{optional}}

\item{\code{statementSubType}}{character \link{optional}}

\item{\code{accessionID}}{integer \link{optional}}

\item{\code{abbreviatedStatementType}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XBRLstatement-new}{\code{XBRLstatement$new()}}
\item \href{#method-XBRLstatement-toJSON}{\code{XBRLstatement$toJSON()}}
\item \href{#method-XBRLstatement-fromJSON}{\code{XBRLstatement$fromJSON()}}
\item \href{#method-XBRLstatement-toJSONString}{\code{XBRLstatement$toJSONString()}}
\item \href{#method-XBRLstatement-fromJSONString}{\code{XBRLstatement$fromJSONString()}}
\item \href{#method-XBRLstatement-validateJSON}{\code{XBRLstatement$validateJSON()}}
\item \href{#method-XBRLstatement-toString}{\code{XBRLstatement$toString()}}
\item \href{#method-XBRLstatement-isValid}{\code{XBRLstatement$isValid()}}
\item \href{#method-XBRLstatement-getInvalidFields}{\code{XBRLstatement$getInvalidFields()}}
\item \href{#method-XBRLstatement-print}{\code{XBRLstatement$print()}}
\item \href{#method-XBRLstatement-clone}{\code{XBRLstatement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLstatement-new"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLstatement-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new XBRLstatement class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLstatement$new(
  name = NULL,
  statementType = NULL,
  statementSubType = NULL,
  accessionID = NULL,
  abbreviatedStatementType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{statementType}}{statementType}

\item{\code{statementSubType}}{statementSubType}

\item{\code{accessionID}}{accessionID}

\item{\code{abbreviatedStatementType}}{abbreviatedStatementType}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLstatement-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLstatement-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLstatement$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
XBRLstatement in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLstatement-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLstatement-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of XBRLstatement
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLstatement$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of XBRLstatement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLstatement-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLstatement-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLstatement$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
XBRLstatement in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLstatement-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLstatement-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of XBRLstatement
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLstatement$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of XBRLstatement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLstatement-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLstatement-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to XBRLstatement and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLstatement$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLstatement-toString"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLstatement-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLstatement$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of XBRLstatement
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLstatement-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLstatement-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLstatement$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLstatement-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLstatement-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLstatement$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLstatement-print"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLstatement-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLstatement$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLstatement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLstatement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLstatement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
