% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_tag_query_parameters.R
\docType{class}
\name{XBRLTagQueryParameters}
\alias{XBRLTagQueryParameters}
\title{XBRLTagQueryParameters}
\format{
An \code{R6Class} generator object
}
\description{
XBRLTagQueryParameters Class

XBRLTagQueryParameters Class
}
\details{
Create a new XBRLTagQueryParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tags}}{list(character) \link{optional}}

\item{\code{metrics}}{list(character) \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XBRLTagQueryParameters-new}{\code{XBRLTagQueryParameters$new()}}
\item \href{#method-XBRLTagQueryParameters-toJSON}{\code{XBRLTagQueryParameters$toJSON()}}
\item \href{#method-XBRLTagQueryParameters-fromJSON}{\code{XBRLTagQueryParameters$fromJSON()}}
\item \href{#method-XBRLTagQueryParameters-toJSONString}{\code{XBRLTagQueryParameters$toJSONString()}}
\item \href{#method-XBRLTagQueryParameters-fromJSONString}{\code{XBRLTagQueryParameters$fromJSONString()}}
\item \href{#method-XBRLTagQueryParameters-validateJSON}{\code{XBRLTagQueryParameters$validateJSON()}}
\item \href{#method-XBRLTagQueryParameters-toString}{\code{XBRLTagQueryParameters$toString()}}
\item \href{#method-XBRLTagQueryParameters-isValid}{\code{XBRLTagQueryParameters$isValid()}}
\item \href{#method-XBRLTagQueryParameters-getInvalidFields}{\code{XBRLTagQueryParameters$getInvalidFields()}}
\item \href{#method-XBRLTagQueryParameters-print}{\code{XBRLTagQueryParameters$print()}}
\item \href{#method-XBRLTagQueryParameters-clone}{\code{XBRLTagQueryParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagQueryParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagQueryParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new XBRLTagQueryParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagQueryParameters$new(tags = NULL, metrics = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tags}}{tags}

\item{\code{metrics}}{metrics}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagQueryParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagQueryParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagQueryParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
XBRLTagQueryParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagQueryParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagQueryParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of XBRLTagQueryParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagQueryParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of XBRLTagQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagQueryParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagQueryParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagQueryParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
XBRLTagQueryParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagQueryParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagQueryParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of XBRLTagQueryParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagQueryParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of XBRLTagQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagQueryParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagQueryParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to XBRLTagQueryParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagQueryParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagQueryParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagQueryParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagQueryParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of XBRLTagQueryParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagQueryParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagQueryParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagQueryParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagQueryParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagQueryParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagQueryParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagQueryParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagQueryParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagQueryParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagQueryParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagQueryParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagQueryParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
