% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_tag_search_parameters.R
\docType{class}
\name{XBRLTagSearchParameters}
\alias{XBRLTagSearchParameters}
\title{XBRLTagSearchParameters}
\format{
An \code{R6Class} generator object
}
\description{
XBRLTagSearchParameters Class

XBRLTagSearchParameters Class
}
\details{
Create a new XBRLTagSearchParameters
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pageParameters}}{\link{XBRLTagQueryParameters} \link{optional}}

\item{\code{companiesParameters}}{\link{CompaniesParameters} \link{optional}}

\item{\code{periodParameters}}{\link{PeriodParameters} \link{optional}}

\item{\code{metricSearchCriteria}}{\link{MetricSearchCriteria} \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XBRLTagSearchParameters-new}{\code{XBRLTagSearchParameters$new()}}
\item \href{#method-XBRLTagSearchParameters-toJSON}{\code{XBRLTagSearchParameters$toJSON()}}
\item \href{#method-XBRLTagSearchParameters-fromJSON}{\code{XBRLTagSearchParameters$fromJSON()}}
\item \href{#method-XBRLTagSearchParameters-toJSONString}{\code{XBRLTagSearchParameters$toJSONString()}}
\item \href{#method-XBRLTagSearchParameters-fromJSONString}{\code{XBRLTagSearchParameters$fromJSONString()}}
\item \href{#method-XBRLTagSearchParameters-validateJSON}{\code{XBRLTagSearchParameters$validateJSON()}}
\item \href{#method-XBRLTagSearchParameters-toString}{\code{XBRLTagSearchParameters$toString()}}
\item \href{#method-XBRLTagSearchParameters-isValid}{\code{XBRLTagSearchParameters$isValid()}}
\item \href{#method-XBRLTagSearchParameters-getInvalidFields}{\code{XBRLTagSearchParameters$getInvalidFields()}}
\item \href{#method-XBRLTagSearchParameters-print}{\code{XBRLTagSearchParameters$print()}}
\item \href{#method-XBRLTagSearchParameters-clone}{\code{XBRLTagSearchParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagSearchParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagSearchParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new XBRLTagSearchParameters class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagSearchParameters$new(
  pageParameters = NULL,
  companiesParameters = NULL,
  periodParameters = NULL,
  metricSearchCriteria = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pageParameters}}{pageParameters}

\item{\code{companiesParameters}}{companiesParameters}

\item{\code{periodParameters}}{periodParameters}

\item{\code{metricSearchCriteria}}{metricSearchCriteria}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagSearchParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagSearchParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagSearchParameters$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
XBRLTagSearchParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagSearchParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagSearchParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of XBRLTagSearchParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagSearchParameters$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of XBRLTagSearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagSearchParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagSearchParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagSearchParameters$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
XBRLTagSearchParameters in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagSearchParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagSearchParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of XBRLTagSearchParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagSearchParameters$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of XBRLTagSearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagSearchParameters-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagSearchParameters-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to XBRLTagSearchParameters and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagSearchParameters$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagSearchParameters-toString"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagSearchParameters-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagSearchParameters$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of XBRLTagSearchParameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagSearchParameters-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagSearchParameters-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagSearchParameters$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagSearchParameters-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagSearchParameters-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagSearchParameters$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagSearchParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagSearchParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagSearchParameters$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XBRLTagSearchParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XBRLTagSearchParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XBRLTagSearchParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
